calculate_footstats: Feature statistics of building footprints

Calculate groups of metrics for building footprint datasets

calculate_footstats(
  X,
  zone = NULL,
  what = "all",
  how = NULL,
  controlZone = list(zoneName = "zoneID", method = "centroid"),
  controlUnits = list(areaUnit = "m^2", perimUnit = "m", distUnit = "m"),
  controlDist = list(maxSearch = 100, method = "centroid", unit =
    controlUnits$distUnit),
  filter = list(minArea = NULL, maxArea = NULL),
  verbose = TRUE
)

# S3 method for sf
calculate_footstats(
  X,
  zone = NULL,
  what = "all",
  how = NULL,
  controlZone = list(zoneName = "zoneID", method = "centroid"),
  controlUnits = list(areaUnit = "m^2", perimUnit = "m", distUnit = "m"),
  controlDist = list(maxSearch = 100, method = "centroid", unit =
    controlUnits$distUnit),
  filter = list(minArea = NULL, maxArea = NULL),
  verbose = TRUE
)

# S3 method for sfc
calculate_footstats(
  X,
  zone = NULL,
  what = "all",
  how = NULL,
  controlZone = list(zoneName = "zoneID", method = "centroid"),
  controlUnits = list(areaUnit = "m^2", perimUnit = "m", distUnit = "m"),
  controlDist = list(maxSearch = 100, method = "centroid", unit =
    controlUnits$distUnit),
  filter = list(minArea = NULL, maxArea = NULL),
  verbose = TRUE
)

# S3 method for sp
calculate_footstats(
  X,
  zone = NULL,
  what = "all",
  how = NULL,
  controlZone = list(zoneName = "zoneID", method = "centroid"),
  controlUnits = list(areaUnit = "m^2", perimUnit = "m", distUnit = "m"),
  controlDist = list(maxSearch = 100, method = "centroid", unit =
    controlUnits$distUnit),
  filter = list(minArea = NULL, maxArea = NULL),
  verbose = TRUE
)

# S3 method for character
calculate_footstats(
  X,
  zone = NULL,
  what = "all",
  how = NULL,
  controlZone = list(zoneName = "zoneID", method = "centroid"),
  controlUnits = list(areaUnit = "m^2", perimUnit = "m", distUnit = "m"),
  controlDist = list(maxSearch = 100, method = "centroid", unit =
    controlUnits$distUnit),
  filter = list(minArea = NULL, maxArea = NULL),
  verbose = TRUE
)

Arguments

X
object with building footprint polygons. This argument can take multiple spatial types, including sf and sp, or a filepath string to a file.

zone
A spatial polygon file defining areas. Or a string identifying a column within X which provides a zonal index for summarising values. Alternatively a vector of indices (with length(index)==nrow(X)) can be provided. If omitted, all observations in X are assumed to be within one zone.

what
list of strings naming the columns or built-in geometry measures to calculate for each footprint. Other options include 'all' or 'nodist' to calculate all available characteristics and all except nearest-neighbour distance metrics.

how
list of strings naming functions to be used to calculate summary statistics. The functions can be built-in functions (e.g. "mean","sd"), or user-defined function names.

controlZone
(optional) named list. Setting controls passed on to zonalIndex. Elements can include zoneName and method.

controlUnits
(optional) named list. Elements can include areaUnit, perimUnit, and distUnit. The values for these items should be strings that can be coerced into a units object.

controlDist
(optional) named list to override default options for distance calculations. Elements can include maxSearch and method. Ignored if metrics does not include a distance calculation. See fs_nndist.

filter
(optional) named list with minArea and maxArea. These are numeric values to filter footprints prior to processing. Default values are NULL and do not filter any records.

verbose
logical. Should progress messages be printed. Default TRUE.

Value

a data.table with an index column identifying the zones and named columns for each footprint summary statistic.

Details

calculate_footstats

calculate_footstats is a wrapper function combining several internal functions from foot. It can calculate various geometric measures for each footprint polygon, including area, perimeter, compactness, shape, angle of rotation, and nearest neighbour distance. To find the list of built-in characteristics and summary metrics, use list_fs().

The what and how arguments are lists specifying the characteristics and the summary metrics to calculated, respectively. Each "how" function will be applied to each "what" characteristic. To apply a metric to only a subset of characteristics, nested lists, with groups of characteristics and functions can be supplied. See examples.

The control arguments are sets of lists with named arguments that pass on these parameters to other foot functions. These are optional.

See also

Examples

data("kampala", package="foot")
buildings <- kampala$buildings
adminzones <- kampala$adminZones

# no summary statistics, just geometry calculations
calculate_footstats(buildings, adminzones, what=list("area","perimeter"))

#> Selecting metrics 
#> Setting control values. 
#> Creating zonal index 
#> Pre-calculating areas 

#> Linking to GEOS 3.8.0, GDAL 3.0.4, PROJ 6.3.1

#> WARNING: different compile-time and runtime versions for GEOS found:

#> Linked against: 3.8.0-CAPI-1.13.1  compiled against: 3.8.1-CAPI-1.13.3

#> It is probably a good idea to reinstall sf, and maybe rgeos and rgdal too

#> Pre-calculating perimeters 
#> No summary functions found, returning metrics. 
#> 

#>                   area     perimeter
#>    1:  559.24140 [m^2] 121.55839 [m]
#>    2:   56.43618 [m^2]  31.20118 [m]
#>    3:   55.84536 [m^2]  34.11899 [m]
#>    4: 2293.58642 [m^2] 220.13423 [m]
#>    5:  229.26511 [m^2]  60.57280 [m]
#>   ---                               
#> 4836:   75.50259 [m^2]  35.91359 [m]
#> 4837:   53.60515 [m^2]  31.22369 [m]
#> 4838:   69.56970 [m^2]  33.72564 [m]
#> 4839:  376.08518 [m^2]  99.95716 [m]
#> 4840:  164.93268 [m^2]  53.28785 [m]

# average building footprint area
calculate_footstats(buildings, 
                    zone=adminzones, 
                    what="area", how="mean")

#> Selecting metrics 
#> Setting control values. 
#> Creating zonal index 
#> Pre-calculating areas 
#> 
#> Calculating  1  metrics ... 
#>    area mean  
#> Finished calculating metrics. 

#>     zoneID        area_mean
#>  1:      1  402.59838 [m^2]
#>  2:      2  211.05341 [m^2]
#>  3:      3  525.07469 [m^2]
#>  4:      4  555.09314 [m^2]
#>  5:      5  568.71544 [m^2]
#>  6:      6 1021.95294 [m^2]
#>  7:      7 2027.70402 [m^2]
#>  8:      8  908.24803 [m^2]
#>  9:      9  181.68220 [m^2]
#> 10:     10  251.99721 [m^2]
#> 11:     11  220.11253 [m^2]
#> 12:     12   86.38345 [m^2]
#> 13:     13  661.61890 [m^2]
#> 14:     14   81.85444 [m^2]
#> 15:     15   80.78532 [m^2]
#> 16:     16   68.75161 [m^2]
#> 17:     17  139.17524 [m^2]
#> 18:     18   87.01754 [m^2]
#> 19:     19  237.07627 [m^2]
#> 20:     20  199.39362 [m^2]
#> 21:     21  244.09906 [m^2]
#> 22:     22  192.39558 [m^2]
#> 23:     23  771.35501 [m^2]
#> 24:     24  219.27229 [m^2]
#> 25:     25  506.22155 [m^2]
#> 26:     26 1164.40867 [m^2]
#> 27:     27  191.05557 [m^2]
#> 28:     29  238.76025 [m^2]
#> 29:     30  203.68136 [m^2]
#> 30:     31  142.13649 [m^2]
#> 31:     32  272.76984 [m^2]
#> 32:     33  418.66391 [m^2]
#> 33:     34  176.81732 [m^2]
#>     zoneID        area_mean

                    
# calculate multiple metrics - nested lists to group arguments
calculate_footstats(buildings, adminzones, 
                    what=list(list("area"), list("perimeter")), 
                    how=list(list("mean","sum"), list("sd","cv")))

#> Selecting metrics 
#> Setting control values. 
#> Creating zonal index 
#> Pre-calculating areas 
#> Pre-calculating perimeters 
#> 
#> Calculating  4  metrics ... 
#>    area mean  
#>    area sum  
#>    perimeter sd  
#>    perimeter cv  
#> Finished calculating metrics. 

#>     zoneID        area_mean          area_sum perimeter_sd perimeter_cv
#>  1:      1  402.59838 [m^2]   8454.5661 [m^2]    55.665551    0.6561304
#>  2:      2  211.05341 [m^2]  29758.5314 [m^2]    44.587096    0.7313348
#>  3:      3  525.07469 [m^2]  25728.6600 [m^2]    88.397185    0.9811076
#>  4:      4  555.09314 [m^2]  63835.7113 [m^2]    88.872380    1.0600954
#>  5:      5  568.71544 [m^2]  26729.6259 [m^2]    88.251084    0.9581201
#>  6:      6 1021.95294 [m^2]  35768.3531 [m^2]   105.385333    0.9616081
#>  7:      7 2027.70402 [m^2]  28387.8563 [m^2]   170.141093    0.9286094
#>  8:      8  908.24803 [m^2]  44504.1537 [m^2]   109.992947    0.8489816
#>  9:      9  181.68220 [m^2]  48145.7825 [m^2]    54.922117    0.9246860
#> 10:     10  251.99721 [m^2]   7559.9164 [m^2]    75.861771    1.0900293
#> 11:     11  220.11253 [m^2]  24212.3780 [m^2]    61.492800    0.8815988
#> 12:     12   86.38345 [m^2]   3109.8042 [m^2]    28.289807    0.7257624
#> 13:     13  661.61890 [m^2]   2646.4756 [m^2]   139.514757    1.3165558
#> 14:     14   81.85444 [m^2]   6712.0641 [m^2]     7.755949    0.2059276
#> 15:     15   80.78532 [m^2]  19226.9062 [m^2]    18.498145    0.5073846
#> 16:     16   68.75161 [m^2]   9900.2318 [m^2]    10.681959    0.3095971
#> 17:     17  139.17524 [m^2]  14752.5754 [m^2]    27.254588    0.5541695
#> 18:     18   87.01754 [m^2]  36895.4355 [m^2]    26.337404    0.7093738
#> 19:     19  237.07627 [m^2]  96490.0399 [m^2]    26.881426    0.4151437
#> 20:     20  199.39362 [m^2]  80355.6287 [m^2]    30.741879    0.5257108
#> 21:     21  244.09906 [m^2]  18063.3304 [m^2]    46.479887    0.7810948
#> 22:     22  192.39558 [m^2]  10196.9657 [m^2]    12.929421    0.2238327
#> 23:     23  771.35501 [m^2] 132673.0618 [m^2]   113.778927    1.1205392
#> 24:     24  219.27229 [m^2]  61834.7854 [m^2]    33.388278    0.5317242
#> 25:     25  506.22155 [m^2]  11643.0956 [m^2]    84.106005    0.8940386
#> 26:     26 1164.40867 [m^2] 150208.7190 [m^2]   146.936669    1.0767096
#> 27:     27  191.05557 [m^2]  33816.8355 [m^2]    25.259149    0.4462858
#> 28:     29  238.76025 [m^2]  37007.8380 [m^2]    63.489927    1.0713172
#> 29:     30  203.68136 [m^2]  15276.1021 [m^2]    33.438265    0.5426390
#> 30:     31  142.13649 [m^2]    142.1365 [m^2]           NA           NA
#> 31:     32  272.76984 [m^2]  35732.8487 [m^2]    45.074544    0.6469339
#> 32:     33  418.66391 [m^2]   7117.2864 [m^2]    55.919761    0.6104610
#> 33:     34  176.81732 [m^2] 146935.1905 [m^2]    42.933274    0.7474269
#>     zoneID        area_mean          area_sum perimeter_sd perimeter_cv