Compactness index

Calculates an approximate measure of shape "compactness".

Calculates a measure of shape or "roundness".

fs_compact(X)

fs_shape(X, unit = NULL)

Arguments

X
polygons of building footprints of type sf.

unit
string indicating unit of measure. Passed to units::set_units.

Source

Polsby, Daniel D., and Robert D. Popper. 1991. “The Third Criterion: Compactness as a procedural safeguard against partisan gerrymandering.” Yale Law & Policy Review 9 (2): 301–353.

Value

numeric vector of compactness values ranging from 0 to 1 for each item in X.

numeric vector with shape index values ranging from 0 to 1 or each item in X.

Details

The compactness measure is the Polsby-Popper test based on the perimeter and area of each footprint.

$$ PP_i = \frac{4$\pi$ * A_z}{P_i^2} $$, where P and A are the perimeter and area of footprint 'i', respectively. Values closer to 1 indicate more compact shapes.

The shape index is calculated as the ratio of footprint polygon's area to the area of minimum bounding circle. The mbc is calculated using fs_mbc and lwgeom. Values closer 1 suggest more rounded shapes.

The function first looks for pre-calculated values of area in a field called fs_area. If not present, the areas are calculated in m^2.

Examples

data("kampala", package="foot")
buildings = kampala$buildings

fs_shape(buildings)

#>    [1] 0.5732678 0.3936447 0.5940971 0.4757789 0.5279643 0.4594504 0.5520443
#>    [8] 0.6186290 0.6146992 0.3191904 0.5831906 0.3734735 0.6199272 0.6284222
#>   [15] 0.5296344 0.5650631 0.6277433 0.3951915 0.4613216 0.4603268 0.5462152
#>   [22] 0.5498982 0.3694682 0.6364904 0.6306219 0.6126685 0.5760368 0.4579810
#>   [29] 0.4540753 0.5232659 0.5624978 0.6091826 0.4467311 0.4729081 0.6155498
#>   [36] 0.5747624 0.4689331 0.2900026 0.5714816 0.6252113 0.2084647 0.5164598
#>   [43] 0.5895023 0.4781961 0.6219324 0.4698386 0.4870994 0.6131092 0.4987130
#>   [50] 0.5717635 0.6246024 0.6336820 0.3113283 0.6368502 0.5620988 0.5332064
#>   [57] 0.5861844 0.5753725 0.5172887 0.3914588 0.5730956 0.6212242 0.5352548
#>   [64] 0.6166964 0.5826929 0.6194326 0.6346224 0.6360538 0.6337825 0.3677919
#>   [71] 0.6074873 0.4088533 0.6211853 0.3858857 0.6314269 0.6312969 0.5383333
#>   [78] 0.5501454 0.5237340 0.5971236 0.6362928 0.4894177 0.5507648 0.5254036
#>   [85] 0.6256829 0.3799601 0.4169084 0.6101312 0.4104715 0.6279283 0.5463382
#>   [92] 0.5938875 0.6168178 0.5788180 0.6059259 0.6118937 0.5606798 0.3868823
#>   [99] 0.4737259 0.5770827 0.3220908 0.5360759 0.4319531 0.5599976 0.6290193
#>  [106] 0.5668814 0.6205110 0.6293345 0.6272624 0.6122671 0.5587328 0.5049883
#>  [113] 0.3051793 0.6009442 0.5759303 0.5329554 0.4953879 0.4195626 0.3488801
#>  [120] 0.3713825 0.6293931 0.4079712 0.6342394 0.6353556 0.5344724 0.6166059
#>  [127] 0.6145827 0.4787557 0.6159380 0.4435303 0.5680683 0.4134564 0.5604033
#>  [134] 0.6223554 0.6222141 0.4666822 0.4104654 0.6251521 0.5751683 0.3029485
#>  [141] 0.6270020 0.4170419 0.5508245 0.5442411 0.3209490 0.5772932 0.3654246
#>  [148] 0.4395289 0.5472500 0.6357580 0.5314029 0.6346223 0.6219428 0.5127167
#>  [155] 0.4165804 0.5605343 0.6110701 0.5840333 0.6135174 0.2914852 0.5835351
#>  [162] 0.4573074 0.4907641 0.6207789 0.4516018 0.6286748 0.4449221 0.6152143
#>  [169] 0.3954646 0.4889778 0.4554410 0.5491443 0.6173362 0.6249906 0.5291472
#>  [176] 0.4382498 0.5110419 0.5807753 0.3082652 0.5924815 0.6318612 0.5455833
#>  [183] 0.2144941 0.5903941 0.6158345 0.5992872 0.6310110 0.5682653 0.6028650
#>  [190] 0.5014108 0.3321919 0.3237417 0.5741066 0.3406314 0.4787621 0.5116576
#>  [197] 0.6266792 0.5881813 0.6302426 0.4981359 0.5232290 0.6051377 0.6286266
#>  [204] 0.5867089 0.6312256 0.2380819 0.5103478 0.3343465 0.4957880 0.6345263
#>  [211] 0.6073142 0.5310449 0.4946763 0.5352302 0.3151704 0.6353934 0.5233142
#>  [218] 0.6089194 0.6350120 0.4778251 0.3352737 0.3249774 0.5511596 0.2976069
#>  [225] 0.3716694 0.4085868 0.4476640 0.4888033 0.2567916 0.5781290 0.5863180
#>  [232] 0.6333642 0.6349956 0.4497027 0.4184146 0.6126086 0.3763936 0.3140236
#>  [239] 0.5297468 0.4694345 0.6081450 0.5412178 0.6154856 0.6196409 0.6266618
#>  [246] 0.6112886 0.4553351 0.5171947 0.5201907 0.5258431 0.4655979 0.2819886
#>  [253] 0.6282462 0.4215686 0.5548060 0.6110211 0.2804562 0.6155576 0.6321669
#>  [260] 0.5912455 0.6334694 0.5731068 0.4794715 0.5073525 0.3337234 0.4608594
#>  [267] 0.5734344 0.5899185 0.6132101 0.5943561 0.5992756 0.5602494 0.6331642
#>  [274] 0.4283730 0.3656978 0.4413582 0.4907641 0.3340619 0.6220353 0.3498178
#>  [281] 0.5828180 0.3540712 0.5617178 0.6171665 0.5951295 0.4344918 0.6234890
#>  [288] 0.2355929 0.2820776 0.4109798 0.5669007 0.4878995 0.5650191 0.3937041
#>  [295] 0.4946734 0.6162411 0.5884730 0.6117822 0.3905867 0.5665231 0.5751597
#>  [302] 0.5783630 0.5102864 0.5063289 0.6004538 0.4558801 0.5164325 0.5681197
#>  [309] 0.4237664 0.6190051 0.4553184 0.5952785 0.6265721 0.5135872 0.6154367
#>  [316] 0.4490018 0.5627864 0.4934252 0.6073142 0.5964838 0.6264696 0.4959897
#>  [323] 0.6108065 0.2445629 0.5032890 0.4988642 0.5054230 0.5236548 0.5567676
#>  [330] 0.4868173 0.6032521 0.5343490 0.4993181 0.6315335 0.4439326 0.2899944
#>  [337] 0.5459235 0.5503130 0.6328835 0.5254626 0.4907641 0.4871390 0.5274040
#>  [344] 0.5160851 0.4560084 0.5511351 0.6254119 0.4514212 0.4867921 0.5775540
#>  [351] 0.5380660 0.4420307 0.5426239 0.5590389 0.3600089 0.5661080 0.4993531
#>  [358] 0.6310840 0.6109732 0.5393442 0.6211657 0.5030186 0.5828334 0.4349429
#>  [365] 0.5177286 0.4375615 0.3364812 0.5769662 0.6242349 0.5030205 0.2939522
#>  [372] 0.2919351 0.5378013 0.3232712 0.6126475 0.6342680 0.4371584 0.4857196
#>  [379] 0.3250053 0.3841080 0.5657426 0.3640386 0.5130931 0.4707294 0.6341424
#>  [386] 0.5789662 0.6197203 0.6149261 0.4034005 0.6259741 0.6294157 0.4701928
#>  [393] 0.6332122 0.6280027 0.5443159 0.4723346 0.5167601 0.5643455 0.2218265
#>  [400] 0.5186334 0.3003906 0.3438613 0.3851202 0.4712818 0.4408044 0.5192705
#>  [407] 0.6129725 0.4977092 0.5078557 0.4621830 0.6264258 0.6218547 0.6364898
#>  [414] 0.5708970 0.4199440 0.5343371 0.5754441 0.6286238 0.5071243 0.2731176
#>  [421] 0.3588341 0.3524820 0.5265317 0.4690751 0.3075076 0.2981466 0.6040803
#>  [428] 0.5376357 0.6166603 0.6089667 0.3953501 0.4411254 0.3779296 0.2095703
#>  [435] 0.5950563 0.6325632 0.4739805 0.4341733 0.4031459 0.6271121 0.4640386
#>  [442] 0.4204684 0.5244979 0.5896837 0.6341160 0.5777366 0.5810100 0.6310727
#>  [449] 0.5443588 0.4898007 0.3487655 0.3660179 0.3422738 0.4355351 0.6234071
#>  [456] 0.5988555 0.6054012 0.6272093 0.6317466 0.4240556 0.4319854 0.5344765
#>  [463] 0.4847704 0.3086354 0.4863591 0.4843589 0.5501972 0.5829448 0.3565755
#>  [470] 0.6207805 0.4821481 0.5681095 0.6280396 0.2481637 0.4765680 0.6010243
#>  [477] 0.5234883 0.5851545 0.6194843 0.5694788 0.3602561 0.4437520 0.4325012
#>  [484] 0.3363199 0.4177446 0.6306646 0.5204182 0.6263994 0.3763351 0.6310638
#>  [491] 0.2116968 0.4607550 0.6030426 0.6093215 0.6217589 0.5073896 0.4223739
#>  [498] 0.6080276 0.5180114 0.4505511 0.6281866 0.4015600 0.3741296 0.5226196
#>  [505] 0.3727357 0.6316943 0.6341449 0.3506984 0.5189132 0.6182346 0.4860988
#>  [512] 0.5230493 0.5597450 0.3518304 0.5665207 0.6119530 0.6236504 0.3865508
#>  [519] 0.6312432 0.6001376 0.5095286 0.5662659 0.4521174 0.6009245 0.5902573
#>  [526] 0.6366105 0.5133393 0.6335508 0.4319895 0.5544088 0.5504796 0.2173761
#>  [533] 0.4464664 0.6174927 0.6184820 0.5832016 0.6044961 0.6021951 0.2504613
#>  [540] 0.5945696 0.4288784 0.3144497 0.5785803 0.5621871 0.6273353 0.6226984
#>  [547] 0.6332083 0.3712081 0.5006005 0.5906674 0.6278848 0.3681065 0.5976103
#>  [554] 0.5413114 0.3895227 0.5980451 0.6107412 0.6299797 0.4624702 0.5142206
#>  [561] 0.4598335 0.5736644 0.5492271 0.6167371 0.6161169 0.4927962 0.6282860
#>  [568] 0.5953843 0.5759181 0.2594583 0.2372441 0.2862208 0.6273304 0.3855587
#>  [575] 0.6058378 0.6031919 0.3428940 0.6252067 0.5812912 0.6342819 0.4705475
#>  [582] 0.3370678 0.4906452 0.5568020 0.6192487 0.5896647 0.5642280 0.5369756
#>  [589] 0.5756477 0.3033684 0.4692259 0.4796803 0.6198068 0.5964501 0.5795840
#>  [596] 0.6302064 0.6226917 0.5470346 0.4153691 0.5168656 0.4447555 0.4233265
#>  [603] 0.4452253 0.5692880 0.6335139 0.6301792 0.4647649 0.3793293 0.4061856
#>  [610] 0.5289069 0.6080375 0.6303454 0.3116544 0.5997251 0.6162530 0.6068486
#>  [617] 0.6364212 0.6196749 0.3534939 0.2156253 0.5039801 0.4563235 0.3045338
#>  [624] 0.2421332 0.5513591 0.4534947 0.4758280 0.4551950 0.4427931 0.5528198
#>  [631] 0.5917760 0.6271121 0.5893848 0.5003423 0.4305480 0.4985171 0.4656116
#>  [638] 0.6063199 0.3658697 0.4620063 0.5337521 0.4537624 0.6206490 0.4878101
#>  [645] 0.3942523 0.6226149 0.5676665 0.6343995 0.6191050 0.4423193 0.4037742
#>  [652] 0.5591203 0.5238698 0.3113517 0.3943050 0.6310893 0.1601515 0.4526353
#>  [659] 0.2352559 0.4025777 0.4744580 0.3974717 0.6288183 0.4759928 0.6311699
#>  [666] 0.5691454 0.6137143 0.4463501 0.4937057 0.4166530 0.6013750 0.4656407
#>  [673] 0.4196051 0.5041546 0.3977261 0.2927175 0.5139093 0.6336536 0.6339716
#>  [680] 0.3920874 0.6004326 0.5188116 0.6172447 0.6158111 0.4599246 0.5743458
#>  [687] 0.4396942 0.3240909 0.5201656 0.6255879 0.3728243 0.3581155 0.6348061
#>  [694] 0.5633940 0.4245637 0.5810329 0.5603739 0.6253233 0.6077589 0.5941164
#>  [701] 0.6191412 0.5396765 0.4739473 0.4038055 0.5906006 0.5502822 0.5704221
#>  [708] 0.2693133 0.4381253 0.6244786 0.5972418 0.6360792 0.6184304 0.6301824
#>  [715] 0.3368945 0.4461891 0.6021506 0.5420964 0.5782848 0.6212046 0.3650180
#>  [722] 0.6183569 0.3572184 0.5348791 0.4438406 0.3506319 0.4645440 0.6225982
#>  [729] 0.3159928 0.6141071 0.4974573 0.3637648 0.5259200 0.5316568 0.2017726
#>  [736] 0.6277894 0.5289224 0.4836640 0.5295845 0.6240670 0.4903332 0.5279135
#>  [743] 0.5171387 0.3514787 0.5469756 0.6289618 0.6300875 0.5478360 0.6365043
#>  [750] 0.4316075 0.4477988 0.3861417 0.6235276 0.5244979 0.3555304 0.5199927
#>  [757] 0.5687696 0.4349283 0.4375988 0.6320341 0.5795257 0.4391308 0.6207179
#>  [764] 0.5268333 0.5689056 0.3814242 0.5961299 0.5864451 0.5589906 0.3785462
#>  [771] 0.3316558 0.5324319 0.5791552 0.4950032 0.6310956 0.3960936 0.5487231
#>  [778] 0.6162785 0.5468471 0.6301379 0.5995088 0.6228951 0.5401568 0.5631451
#>  [785] 0.6127452 0.5816183 0.4899225 0.2738832 0.5014003 0.6284462 0.6337495
#>  [792] 0.6048168 0.5918542 0.6292239 0.4592503 0.5718668 0.5635348 0.5854236
#>  [799] 0.6221531 0.5449257 0.5285727 0.5627531 0.5806877 0.3841315 0.6361135
#>  [806] 0.4953879 0.2667857 0.3993133 0.2882159 0.4968650 0.6374092 0.5240190
#>  [813] 0.5362542 0.4707714 0.6339187 0.6041887 0.6267874 0.4756079 0.5976726
#>  [820] 0.6260234 0.4194106 0.6284444 0.6350874 0.6087920 0.6356295 0.5958984
#>  [827] 0.6310165 0.3259878 0.5644550 0.4025122 0.4647859 0.3350220 0.4787997
#>  [834] 0.5147360 0.4605672 0.4060271 0.3340630 0.6292629 0.6265778 0.3932861
#>  [841] 0.3936136 0.6181260 0.2751823 0.5248792 0.6232469 0.5162156 0.5780399
#>  [848] 0.4034049 0.5181426 0.4807423 0.6316878 0.6334854 0.6310600 0.4839599
#>  [855] 0.5750945 0.5370777 0.5093767 0.5120312 0.5233530 0.6354202 0.2962075
#>  [862] 0.5782389 0.5683228 0.4611922 0.5993435 0.6299576 0.3967891 0.5636027
#>  [869] 0.3589969 0.6286656 0.3697711 0.3839317 0.5386236 0.4775033 0.3151845
#>  [876] 0.3848270 0.5199588 0.3576235 0.3771404 0.5188192 0.6331997 0.4791474
#>  [883] 0.6147497 0.5578307 0.6313849 0.6328008 0.6224534 0.5193307 0.4795769
#>  [890] 0.6086296 0.5900339 0.4970858 0.4662972 0.3162165 0.6176069 0.5701666
#>  [897] 0.4242119 0.6240008 0.6042487 0.6096152 0.5053936 0.5363969 0.5835738
#>  [904] 0.3125335 0.5044610 0.5237159 0.4798265 0.5513144 0.6316291 0.3722556
#>  [911] 0.3716827 0.6111690 0.3147571 0.1665517 0.6008455 0.3884745 0.2472650
#>  [918] 0.6297406 0.5196550 0.6048467 0.5976104 0.4102201 0.5669132 0.3150126
#>  [925] 0.4263609 0.5836941 0.4205463 0.4141820 0.6175562 0.6280487 0.5094888
#>  [932] 0.3006627 0.2339601 0.2403166 0.4235082 0.5410471 0.3295207 0.6274545
#>  [939] 0.5640970 0.4055722 0.6021402 0.2571944 0.5019383 0.6317515 0.4388103
#>  [946] 0.6352488 0.6037102 0.4928956 0.4744770 0.4112467 0.6221249 0.5744561
#>  [953] 0.5196747 0.5177326 0.4725303 0.5587120 0.5585543 0.1482683 0.5471281
#>  [960] 0.5538262 0.6132562 0.5714376 0.4901128 0.4060361 0.5639522 0.6348973
#>  [967] 0.4908303 0.6177693 0.3039920 0.5438751 0.4519373 0.6351847 0.6227524
#>  [974] 0.5211898 0.3077041 0.5175892 0.5463324 0.3954047 0.3008286 0.6282458
#>  [981] 0.3172420 0.5813039 0.4692552 0.5616511 0.4113245 0.6319062 0.4315112
#>  [988] 0.6232947 0.3274400 0.4196014 0.5112491 0.5636123 0.5357312 0.5164997
#>  [995] 0.4347852 0.5814913 0.5295936 0.6302913 0.5095286 0.4470930 0.6220671
#> [1002] 0.6031639 0.5949173 0.6332382 0.5989499 0.4453557 0.6328397 0.5651983
#> [1009] 0.6081824 0.5669770 0.5472538 0.5261338 0.5440786 0.6110058 0.5082158
#> [1016] 0.4718330 0.5522994 0.5560955 0.4491827 0.5057781 0.6105555 0.4780435
#> [1023] 0.6116347 0.6038173 0.3424303 0.3603820 0.2130853 0.3988120 0.2610077
#> [1030] 0.3814500 0.3731426 0.6180981 0.3364812 0.6106710 0.5128443 0.6151949
#> [1037] 0.5981455 0.5689522 0.6093411 0.3599769 0.5463804 0.6335957 0.4059629
#> [1044] 0.5331887 0.4484592 0.5590001 0.6327797 0.5436612 0.5495525 0.5807642
#> [1051] 0.6128103 0.2821544 0.5573910 0.4761120 0.5243067 0.4863942 0.6291945
#> [1058] 0.5738118 0.5922986 0.4232090 0.6236660 0.4973890 0.4038656 0.5869737
#> [1065] 0.6191795 0.6176245 0.4489892 0.5195963 0.5210007 0.5380162 0.6338625
#> [1072] 0.6290477 0.4816915 0.4699936 0.3456573 0.4313712 0.5583242 0.3774122
#> [1079] 0.5187075 0.3385636 0.5766080 0.4384961 0.4563239 0.6331066 0.3736196
#> [1086] 0.5947685 0.4947165 0.5546529 0.4455735 0.5281412 0.6302644 0.4240166
#> [1093] 0.6301045 0.4300325 0.6331038 0.5411009 0.4835476 0.4245399 0.5643301
#> [1100] 0.4911514 0.3679446 0.6028071 0.5493202 0.5818484 0.4996365 0.5383893
#> [1107] 0.6296581 0.5610419 0.4290309 0.5697884 0.4876327 0.5334301 0.6214042
#> [1114] 0.6242313 0.5279247 0.3855861 0.6086947 0.6281854 0.3786904 0.6059705
#> [1121] 0.2535309 0.6340288 0.5594659 0.6125916 0.4160735 0.5635052 0.4950762
#> [1128] 0.4331312 0.4884424 0.5715263 0.3458089 0.3655646 0.5036874 0.6232772
#> [1135] 0.2499420 0.5653820 0.6155406 0.5865017 0.2647364 0.1980682 0.5651189
#> [1142] 0.6263176 0.5027790 0.4218356 0.3580471 0.6171856 0.5077206 0.5458813
#> [1149] 0.6259834 0.6297200 0.4418360 0.6227347 0.5706768 0.6196969 0.6354395
#> [1156] 0.4550148 0.5848787 0.4938661 0.5647317 0.4791295 0.4863490 0.6065349
#> [1163] 0.6339053 0.6291435 0.6331583 0.2358555 0.5680447 0.4994312 0.3529504
#> [1170] 0.6146715 0.6317046 0.3084176 0.5801742 0.3492813 0.5661429 0.5810505
#> [1177] 0.4668321 0.5739642 0.4232760 0.5200966 0.6164735 0.6312806 0.5540663
#> [1184] 0.4969295 0.4102588 0.4164427 0.5657183 0.5887777 0.4635430 0.2040928
#> [1191] 0.5754671 0.3152714 0.5703638 0.4391803 0.6312119 0.4716953 0.5158079
#> [1198] 0.6142173 0.4995656 0.6353242 0.6167452 0.5114182 0.4269274 0.6241197
#> [1205] 0.5978878 0.5409567 0.3522446 0.6350451 0.5440237 0.3962578 0.5740443
#> [1212] 0.2733090 0.6220989 0.5441687 0.5600073 0.3422382 0.3836823 0.6247989
#> [1219] 0.5252467 0.4477855 0.4956361 0.6065357 0.3377301 0.5913687 0.5682337
#> [1226] 0.3786343 0.3460020 0.2519237 0.6042218 0.5471968 0.3315590 0.4970383
#> [1233] 0.5816658 0.4974056 0.4294612 0.5675347 0.6202196 0.6343995 0.5662339
#> [1240] 0.3066187 0.5884770 0.5597415 0.4816611 0.5456296 0.5817463 0.2912314
#> [1247] 0.4911649 0.6104210 0.6147439 0.6368701 0.3366291 0.5475173 0.4425569
#> [1254] 0.4999439 0.5678327 0.4872547 0.5613643 0.4713945 0.5598932 0.6357299
#> [1261] 0.4043936 0.6261472 0.4924050 0.6160242 0.4680964 0.6367298 0.5177456
#> [1268] 0.6283798 0.6282157 0.2980257 0.5299253 0.4584693 0.2777925 0.6146277
#> [1275] 0.6178621 0.5350980 0.6334690 0.6077421 0.3503876 0.5926009 0.5645429
#> [1282] 0.4508414 0.6240607 0.5530809 0.5230861 0.6184137 0.4504311 0.5232912
#> [1289] 0.1651927 0.4719408 0.4932094 0.5911104 0.5244371 0.5634945 0.5780051
#> [1296] 0.5169905 0.3957591 0.4517692 0.5402787 0.5437098 0.5212201 0.5487144
#> [1303] 0.4910609 0.5330297 0.4387319 0.4304744 0.3845017 0.4730079 0.5163177
#> [1310] 0.5078073 0.6315817 0.5080042 0.5173037 0.3910029 0.6239493 0.5397496
#> [1317] 0.5832836 0.5870090 0.4363162 0.3533931 0.5290844 0.6120697 0.6217346
#> [1324] 0.5957054 0.6175108 0.6268643 0.3341747 0.3015113 0.6184142 0.5833167
#> [1331] 0.4767627 0.4209834 0.4351852 0.4955233 0.3594633 0.5898050 0.6045344
#> [1338] 0.5635501 0.4628300 0.4604791 0.6024449 0.2893889 0.5749754 0.4549577
#> [1345] 0.6313494 0.3366261 0.5007452 0.5854610 0.6124001 0.5670762 0.6338316
#> [1352] 0.6101632 0.6047588 0.6041393 0.3889396 0.6102490 0.5077942 0.5087929
#> [1359] 0.4704248 0.6181155 0.6110797 0.5851148 0.4507212 0.2697362 0.6137332
#> [1366] 0.6202271 0.4233935 0.6143860 0.5960395 0.5857242 0.4136332 0.6291836
#> [1373] 0.5467663 0.6348779 0.3898129 0.4807697 0.4792792 0.3733324 0.3052003
#> [1380] 0.6277507 0.3236390 0.4026693 0.6101663 0.4783574 0.6293726 0.5298123
#> [1387] 0.4163071 0.4357018 0.4492657 0.6083924 0.4626191 0.6062229 0.4116021
#> [1394] 0.5794278 0.4331861 0.5376289 0.6304009 0.4395190 0.5786042 0.5793441
#> [1401] 0.6304955 0.6315939 0.6288941 0.5577004 0.6259897 0.6159475 0.4838800
#> [1408] 0.5311263 0.5547332 0.5473510 0.4343989 0.6369108 0.5792325 0.6081165
#> [1415] 0.5559821 0.6127275 0.6329253 0.4449176 0.4246888 0.4837832 0.6306047
#> [1422] 0.5853765 0.6231551 0.4503530 0.4414938 0.5782585 0.4457533 0.6002145
#> [1429] 0.3549350 0.4954532 0.4846735 0.4300163 0.6302885 0.6065817 0.4723503
#> [1436] 0.6302916 0.2460225 0.6239239 0.3783979 0.6127085 0.6034636 0.3289884
#> [1443] 0.2807098 0.4588280 0.6363089 0.5725040 0.2593765 0.5299879 0.6221074
#> [1450] 0.5913897 0.6274383 0.3755199 0.5202661 0.5445303 0.3752771 0.5868231
#> [1457] 0.5711646 0.5172069 0.5071114 0.4207700 0.5634904 0.6130070 0.5765700
#> [1464] 0.6151691 0.5095900 0.6139294 0.6308741 0.4339946 0.4129854 0.5888329
#> [1471] 0.4726397 0.5556922 0.3708284 0.2586951 0.3328556 0.2889363 0.5516924
#> [1478] 0.4918336 0.6324149 0.5737925 0.4425483 0.4149613 0.4739835 0.3870572
#> [1485] 0.5987156 0.5463854 0.5333949 0.6172447 0.4433606 0.5063457 0.4399225
#> [1492] 0.5404105 0.5915006 0.6156015 0.5626380 0.5053854 0.6209206 0.4434427
#> [1499] 0.6100143 0.6086036 0.4983854 0.6231710 0.4794268 0.4589487 0.5932644
#> [1506] 0.4110069 0.4176182 0.5350045 0.4465063 0.5387905 0.5327297 0.4038936
#> [1513] 0.5854689 0.5314152 0.3039854 0.3259169 0.6062567 0.6329588 0.6022867
#> [1520] 0.6314045 0.6075574 0.6164929 0.2751976 0.6300714 0.4074886 0.6244832
#> [1527] 0.6278038 0.4949862 0.6320609 0.4979223 0.5801837 0.2875042 0.2758681
#> [1534] 0.6357924 0.6334570 0.5479339 0.6366533 0.3911373 0.4749971 0.5985038
#> [1541] 0.4022466 0.5428025 0.4865893 0.6332587 0.4959774 0.6076425 0.6067611
#> [1548] 0.6024831 0.5056665 0.5700492 0.5236830 0.5953676 0.5667714 0.6158353
#> [1555] 0.5728072 0.5469098 0.4852724 0.4802227 0.5956802 0.5391065 0.6218523
#> [1562] 0.4098597 0.6034625 0.3450601 0.4125250 0.4606102 0.6264729 0.5123337
#> [1569] 0.5218294 0.6136091 0.3912691 0.6279583 0.3259670 0.6117313 0.6012018
#> [1576] 0.6215870 0.3402441 0.5547287 0.5453705 0.3925839 0.1948972 0.4552864
#> [1583] 0.5687390 0.3960978 0.4028870 0.2387630 0.5719250 0.3310889 0.3454837
#> [1590] 0.4356714 0.5909342 0.2985915 0.3560171 0.6339004 0.6282737 0.3376678
#> [1597] 0.5295563 0.6365869 0.4750434 0.5289899 0.5775270 0.3784654 0.4701707
#> [1604] 0.2963073 0.4474271 0.5341743 0.6046494 0.4778616 0.5353820 0.4741632
#> [1611] 0.5044811 0.5436225 0.6362431 0.4612441 0.5707936 0.2933016 0.3550865
#> [1618] 0.5506142 0.3620883 0.4737110 0.5117952 0.6088464 0.4941358 0.6274676
#> [1625] 0.6364398 0.5827809 0.4556713 0.5745127 0.6227192 0.5102415 0.6150063
#> [1632] 0.6054388 0.4983484 0.3636911 0.6101892 0.5860759 0.5232159 0.5146076
#> [1639] 0.4046630 0.4465792 0.6106464 0.2373987 0.6231665 0.4778332 0.3437169
#> [1646] 0.2980782 0.2858225 0.6329166 0.6258147 0.4482009 0.4881869 0.2375925
#> [1653] 0.5862152 0.4355302 0.6359367 0.5663658 0.4073284 0.4629335 0.6251578
#> [1660] 0.6113844 0.5097586 0.5829084 0.5986961 0.3414249 0.5994139 0.6054066
#> [1667] 0.4485667 0.5769240 0.5496487 0.5396145 0.6369108 0.6091952 0.4772048
#> [1674] 0.5656778 0.5475355 0.3847365 0.5097541 0.5805262 0.5000005 0.6151219
#> [1681] 0.6051117 0.3627772 0.5923451 0.5458813 0.2755551 0.6338613 0.5769675
#> [1688] 0.5650213 0.5732197 0.5559324 0.5904277 0.4505185 0.2769967 0.5877175
#> [1695] 0.4490052 0.2869037 0.5444260 0.5830827 0.6336506 0.5229710 0.4628339
#> [1702] 0.4621396 0.5179274 0.5328113 0.5736486 0.3765249 0.3592440 0.6097592
#> [1709] 0.6124910 0.6368409 0.6343077 0.3991913 0.5717661 0.4978006 0.5201080
#> [1716] 0.5749754 0.6348699 0.5738559 0.6106335 0.5943237 0.5071994 0.5888308
#> [1723] 0.5725040 0.6116524 0.6114343 0.4143209 0.5095286 0.6337242 0.5617567
#> [1730] 0.5486863 0.3703925 0.3851718 0.5442044 0.4475834 0.6246317 0.5791729
#> [1737] 0.5500823 0.5284378 0.6260074 0.6004086 0.3268480 0.6314817 0.5902626
#> [1744] 0.5461368 0.6324338 0.3431131 0.5046846 0.6271121 0.5429748 0.4375269
#> [1751] 0.6287368 0.4521991 0.6227444 0.4833830 0.5576857 0.4618901 0.4730532
#> [1758] 0.4448124 0.6328543 0.4489654 0.5182548 0.4367473 0.5303001 0.5706550
#> [1765] 0.5913448 0.4733221 0.4300174 0.6115818 0.4850977 0.4484756 0.6166940
#> [1772] 0.3346982 0.6178925 0.5304774 0.4276999 0.5507008 0.6120061 0.5093961
#> [1779] 0.4163347 0.4488826 0.3450030 0.5383147 0.4497813 0.3932125 0.4223867
#> [1786] 0.4894421 0.6098640 0.5662777 0.5758627 0.2386097 0.4074269 0.5893481
#> [1793] 0.4135050 0.2729651 0.5104412 0.5082560 0.6311431 0.6366433 0.6273364
#> [1800] 0.4180961 0.5421410 0.5041993 0.4513641 0.4824784 0.6299939 0.6244223
#> [1807] 0.4434381 0.3202607 0.6209807 0.4923363 0.6298660 0.2946075 0.3467535
#> [1814] 0.5363157 0.6230108 0.3341934 0.5193226 0.6306047 0.5677405 0.3684988
#> [1821] 0.4261290 0.4772389 0.5520413 0.5648830 0.5018817 0.5434182 0.5048368
#> [1828] 0.5912429 0.2467870 0.6202251 0.3788473 0.6122493 0.3876369 0.6247276
#> [1835] 0.4213392 0.5593824 0.4254219 0.6330387 0.3918313 0.2481672 0.6256294
#> [1842] 0.5159085 0.3867397 0.3804655 0.4711434 0.5737484 0.4909460 0.4964406
#> [1849] 0.5671162 0.5293040 0.5494225 0.6362762 0.6281091 0.6360738 0.3914221
#> [1856] 0.2947310 0.5439916 0.5208425 0.5548876 0.4545841 0.5742374 0.5721508
#> [1863] 0.6263334 0.3505576 0.3236887 0.4613931 0.4331725 0.5715346 0.5788070
#> [1870] 0.5361152 0.3250926 0.3946836 0.2321738 0.6192943 0.4727594 0.5136745
#> [1877] 0.3684589 0.6020790 0.5662977 0.2313513 0.5735677 0.4815608 0.4555893
#> [1884] 0.6316470 0.5962367 0.4251113 0.5779292 0.3568043 0.6308959 0.3727338
#> [1891] 0.5881029 0.5079565 0.3856457 0.2486141 0.5007130 0.5957278 0.6145475
#> [1898] 0.5634596 0.4325276 0.5210305 0.5228861 0.5949877 0.3689049 0.4091106
#> [1905] 0.6204318 0.4848070 0.3127371 0.4355445 0.5146242 0.6084392 0.6049889
#> [1912] 0.1723415 0.6273395 0.4811193 0.4521038 0.6051717 0.5387527 0.6233449
#> [1919] 0.3821215 0.2916642 0.5595188 0.5467283 0.5658871 0.3263173 0.4134409
#> [1926] 0.5523287 0.5619801 0.3995949 0.3634038 0.4799384 0.5501986 0.4902017
#> [1933] 0.6319851 0.4878586 0.5291597 0.2302777 0.2082574 0.2585341 0.4439692
#> [1940] 0.5880756 0.6227423 0.3952460 0.6288237 0.6304991 0.4640874 0.5451298
#> [1947] 0.6354234 0.6339353 0.6341312 0.5361743 0.4993715 0.5440456 0.5289246
#> [1954] 0.4013164 0.5634082 0.3410614 0.6362700 0.5831015 0.4705202 0.4324600
#> [1961] 0.3646536 0.5511286 0.5281046 0.3706768 0.4644786 0.4738055 0.5187858
#> [1968] 0.3116701 0.4117601 0.6251857 0.5095286 0.4580076 0.5848264 0.6061154
#> [1975] 0.5911658 0.6369108 0.6319912 0.4011945 0.5563847 0.3983620 0.3940645
#> [1982] 0.6369108 0.4995196 0.5696296 0.4741152 0.6339797 0.5302817 0.5375587
#> [1989] 0.5505967 0.4429194 0.2524450 0.2321666 0.5868767 0.3314997 0.3145970
#> [1996] 0.4007895 0.5652139 0.4837733 0.5859579 0.4361608 0.5064796 0.6347225
#> [2003] 0.6227970 0.3012076 0.6036935 0.6117175 0.2851746 0.4228367 0.3755302
#> [2010] 0.4732105 0.4589824 0.3569905 0.4433949 0.6322801 0.5692042 0.5738593
#> [2017] 0.4987189 0.4023200 0.3304653 0.6288937 0.3310831 0.4746252 0.6118703
#> [2024] 0.6144049 0.6287483 0.4237664 0.5543768 0.2806593 0.5038793 0.6195130
#> [2031] 0.5377773 0.5928486 0.2932913 0.4103789 0.3109684 0.3133273 0.4967438
#> [2038] 0.4487668 0.6289837 0.5902682 0.5427947 0.4742065 0.5678580 0.6206711
#> [2045] 0.4539284 0.4681978 0.4961868 0.5775473 0.6360614 0.4065393 0.5519893
#> [2052] 0.5570612 0.4942114 0.4860482 0.4822942 0.6126908 0.4541290 0.4225532
#> [2059] 0.5250257 0.6292393 0.6228168 0.4852533 0.6345268 0.5404944 0.3822136
#> [2066] 0.4410313 0.6170762 0.5832168 0.5191228 0.5314275 0.5868588 0.4887061
#> [2073] 0.3515877 0.4828379 0.4778508 0.6319831 0.3463713 0.5272188 0.6314385
#> [2080] 0.3012434 0.3618968 0.5942236 0.4874214 0.6063199 0.6337424 0.6269164
#> [2087] 0.5840594 0.3784777 0.2469337 0.5194992 0.5351785 0.6287510 0.5037141
#> [2094] 0.4016883 0.6189745 0.4021768 0.5624957 0.5776769 0.5191390 0.6130045
#> [2101] 0.5020766 0.5209182 0.5879176 0.3925898 0.6299197 0.6214905 0.6317435
#> [2108] 0.4311937 0.6097725 0.5217221 0.5768687 0.6247758 0.6049417 0.6369108
#> [2115] 0.4807411 0.6301288 0.4628417 0.6308335 0.2622966 0.3483569 0.5818104
#> [2122] 0.4559684 0.6327929 0.5879176 0.3396496 0.6283961 0.3182979 0.5453085
#> [2129] 0.6300495 0.3594732 0.6100904 0.4159903 0.5385423 0.3924541 0.4183152
#> [2136] 0.2893738 0.5954216 0.5115526 0.3707401 0.3571383 0.3661262 0.5935251
#> [2143] 0.4341708 0.5529604 0.5959270 0.5472076 0.5970875 0.4557615 0.1923282
#> [2150] 0.6007132 0.5809354 0.6090225 0.5198453 0.4780333 0.4391805 0.2254765
#> [2157] 0.3841534 0.6233034 0.6339899 0.5696083 0.6010554 0.6282092 0.6079311
#> [2164] 0.6297936 0.5976370 0.3830892 0.4269228 0.6333701 0.5846759 0.5848891
#> [2171] 0.6166316 0.2837896 0.6334742 0.6360418 0.5108533 0.2738298 0.5216354
#> [2178] 0.5288125 0.6083194 0.6213763 0.3753430 0.5932443 0.5392400 0.4066514
#> [2185] 0.6191503 0.6083241 0.5243458 0.5468608 0.5401285 0.5696813 0.4931611
#> [2192] 0.4775666 0.5727676 0.4673134 0.3555739 0.4436333 0.5139742 0.6352602
#> [2199] 0.2805802 0.6049978 0.5547153 0.5263302 0.6065134 0.4961180 0.5934144
#> [2206] 0.5546777 0.6279392 0.3231651 0.3890731 0.3387102 0.6252705 0.5242038
#> [2213] 0.5672343 0.3458794 0.3828954 0.5929018 0.6326243 0.5605499 0.4327071
#> [2220] 0.4457792 0.4326442 0.5262834 0.2624742 0.4461960 0.3638216 0.5951550
#> [2227] 0.4970481 0.4807834 0.4638633 0.5996843 0.5210779 0.5681611 0.3698501
#> [2234] 0.3184646 0.5282760 0.4324582 0.5292722 0.5409324 0.5344604 0.5532370
#> [2241] 0.5310689 0.6239654 0.4450341 0.6177143 0.5333716 0.4597942 0.5821320
#> [2248] 0.6337195 0.5909444 0.4663343 0.5934818 0.5907443 0.4595876 0.5436928
#> [2255] 0.5590376 0.4160702 0.5414793 0.4165463 0.5746668 0.5360711 0.5391386
#> [2262] 0.5743385 0.3439386 0.6123679 0.4694446 0.3076213 0.5676407 0.5869737
#> [2269] 0.4927378 0.3013651 0.4034729 0.3359507 0.2380542 0.3329654 0.5333437
#> [2276] 0.5803847 0.6328271 0.6106666 0.6275879 0.5789452 0.6103577 0.6073281
#> [2283] 0.4507211 0.5681884 0.5194632 0.3372992 0.4741272 0.6156015 0.4906178
#> [2290] 0.5996650 0.4674449 0.3412221 0.5326154 0.3994207 0.5959646 0.5958600
#> [2297] 0.5750081 0.5756829 0.4507846 0.4528728 0.6184320 0.4549859 0.3732254
#> [2304] 0.6189322 0.6142785 0.5932867 0.3837534 0.3790424 0.4530979 0.5761720
#> [2311] 0.3818318 0.3978669 0.4276281 0.5538898 0.5691852 0.5106539 0.4167065
#> [2318] 0.5960124 0.5983022 0.3263297 0.4477550 0.6298217 0.6360168 0.5070041
#> [2325] 0.5634667 0.4686200 0.4893755 0.6312493 0.6197545 0.6116479 0.6030761
#> [2332] 0.5739670 0.5168780 0.6003202 0.3844853 0.3014954 0.6079428 0.4479259
#> [2339] 0.1752431 0.4173691 0.4793522 0.4975025 0.6294938 0.4725288 0.6334565
#> [2346] 0.4411684 0.5636043 0.4505882 0.6369108 0.5455527 0.3703501 0.4668209
#> [2353] 0.5235062 0.4246854 0.3670233 0.4271055 0.3247420 0.6260854 0.3135832
#> [2360] 0.3697525 0.5842197 0.5733172 0.6052645 0.5034475 0.5704055 0.2531565
#> [2367] 0.4801187 0.5992783 0.6280073 0.6300768 0.6178502 0.6004562 0.5301442
#> [2374] 0.5800967 0.4047390 0.5636574 0.5764434 0.6335237 0.6175939 0.6129217
#> [2381] 0.5346284 0.5329871 0.4111991 0.6079688 0.4900682 0.2430982 0.6151547
#> [2388] 0.6267820 0.6329537 0.6345874 0.4023209 0.6326055 0.3661858 0.2973181
#> [2395] 0.4091822 0.5485328 0.5714558 0.5112101 0.4918849 0.2469303 0.5451783
#> [2402] 0.6097637 0.5773852 0.6262693 0.4940098 0.5959119 0.6369108 0.6303783
#> [2409] 0.5357349 0.6146606 0.6283428 0.3402034 0.5277222 0.6186547 0.5520492
#> [2416] 0.5083864 0.5410163 0.4800616 0.3862879 0.6057115 0.6072872 0.5282701
#> [2423] 0.6345835 0.6049347 0.5851877 0.5620542 0.5306930 0.5536286 0.6286050
#> [2430] 0.5013076 0.4954689 0.1296743 0.3568171 0.6274654 0.4435896 0.6186825
#> [2437] 0.5970241 0.4405928 0.4326853 0.6367696 0.3906594 0.4666475 0.5930427
#> [2444] 0.6127243 0.5058578 0.6108012 0.4421005 0.4050401 0.2318117 0.2529767
#> [2451] 0.5041186 0.4249185 0.5908792 0.6234232 0.6205628 0.4525042 0.4885249
#> [2458] 0.4397302 0.5609056 0.5694688 0.6200867 0.5440851 0.5573630 0.4971039
#> [2465] 0.4242272 0.5339349 0.5060003 0.5207749 0.3597484 0.5040109 0.5195801
#> [2472] 0.5211806 0.6230457 0.5631251 0.4887980 0.4557668 0.4425291 0.5136248
#> [2479] 0.4209043 0.5351493 0.5756509 0.4096971 0.5455459 0.6295388 0.5512319
#> [2486] 0.3386097 0.6302064 0.6298662 0.2325041 0.5842177 0.3681123 0.3910299
#> [2493] 0.5885425 0.5748592 0.3956931 0.5371262 0.4206052 0.6290388 0.6287677
#> [2500] 0.4090968 0.6264637 0.4158932 0.4981844 0.5267908 0.3991386 0.4725953
#> [2507] 0.5985968 0.4068448 0.5753117 0.5356436 0.3930476 0.5387232 0.4377587
#> [2514] 0.5488332 0.5280347 0.4194232 0.6311824 0.4517014 0.5640121 0.6307806
#> [2521] 0.5324713 0.6306038 0.6367781 0.4568365 0.5290649 0.4773241 0.4636560
#> [2528] 0.3657951 0.5639563 0.5068327 0.6146277 0.1183158 0.6338572 0.4793052
#> [2535] 0.3070354 0.2972505 0.5502094 0.5312648 0.2292876 0.4320590 0.3185750
#> [2542] 0.3330567 0.3971827 0.5234300 0.4840239 0.3790703 0.6026778 0.4905183
#> [2549] 0.6369108 0.5817931 0.5532792 0.6187592 0.5570202 0.5747723 0.5783315
#> [2556] 0.4106472 0.5848725 0.6254434 0.3648136 0.3237824 0.5828626 0.4025977
#> [2563] 0.3382024 0.5606056 0.3377969 0.5373758 0.4126384 0.2630703 0.6294944
#> [2570] 0.5789756 0.6185717 0.5421186 0.2769453 0.5320764 0.6283785 0.6106724
#> [2577] 0.4269319 0.5047721 0.3053062 0.4446974 0.3992326 0.5508121 0.3360158
#> [2584] 0.5293794 0.2987329 0.5407733 0.3580120 0.6291118 0.5807258 0.3976633
#> [2591] 0.5988877 0.5674061 0.5850044 0.5811616 0.1986274 0.6302650 0.6046181
#> [2598] 0.5525752 0.6292969 0.5854919 0.6129688 0.6322602 0.5878162 0.5438504
#> [2605] 0.5871575 0.6311881 0.5891719 0.4890592 0.3994753 0.6352113 0.6195691
#> [2612] 0.6292239 0.6180742 0.2172640 0.4014853 0.6332955 0.6282157 0.6162248
#> [2619] 0.4412372 0.4544924 0.6211446 0.5650074 0.3293089 0.6272471 0.5144912
#> [2626] 0.6063345 0.6177196 0.4958107 0.4334473 0.3053451 0.2104625 0.5760921
#> [2633] 0.6192843 0.5476334 0.5596999 0.4361380 0.4739762 0.4572409 0.3928826
#> [2640] 0.4897655 0.4025178 0.6078772 0.5260176 0.5705088 0.6256780 0.5993977
#> [2647] 0.6109831 0.1902332 0.6327196 0.5422001 0.5398596 0.5953136 0.6305568
#> [2654] 0.5404123 0.6350242 0.5187811 0.5180569 0.5486230 0.5282533 0.5231503
#> [2661] 0.6213763 0.6296019 0.6255495 0.5824729 0.4207166 0.3023637 0.4744279
#> [2668] 0.5668813 0.5003127 0.4479076 0.6337244 0.4727554 0.6347187 0.5705904
#> [2675] 0.2652083 0.3968999 0.5466289 0.6308338 0.3987327 0.4836806 0.3874354
#> [2682] 0.5694731 0.3154032 0.4461475 0.6213023 0.3144497 0.6064576 0.6107716
#> [2689] 0.5625431 0.6141071 0.6262876 0.6324886 0.6236263 0.6286893 0.5993499
#> [2696] 0.5578112 0.4176595 0.5044872 0.5861086 0.5987340 0.6093295 0.4844021
#> [2703] 0.5835784 0.3852581 0.5930516 0.6342544 0.5132956 0.3569009 0.4635568
#> [2710] 0.4512323 0.5630165 0.6233494 0.3736551 0.6020090 0.3702003 0.5001737
#> [2717] 0.5905259 0.4954799 0.6136335 0.5565601 0.4523357 0.5847439 0.5774694
#> [2724] 0.5659038 0.5879176 0.4153005 0.5798290 0.4868415 0.3739759 0.2596462
#> [2731] 0.4224091 0.5218188 0.5995738 0.6144620 0.6349954 0.3038382 0.5186097
#> [2738] 0.5947717 0.6030766 0.4949486 0.5769757 0.5508162 0.4489570 0.6078786
#> [2745] 0.6002816 0.5006118 0.4936642 0.5586606 0.5901596 0.6304173 0.5298669
#> [2752] 0.4146583 0.5993613 0.6057459 0.5479110 0.2142894 0.5873376 0.3537759
#> [2759] 0.6287846 0.4442948 0.2947190 0.6296067 0.3539225 0.3481908 0.4009071
#> [2766] 0.5813462 0.5504536 0.6133802 0.3725793 0.3056858 0.4681956 0.4556151
#> [2773] 0.6366618 0.3709998 0.2074071 0.2181195 0.4918071 0.4319284 0.4538834
#> [2780] 0.6037113 0.4146886 0.5743030 0.6295369 0.5377704 0.5368802 0.6208878
#> [2787] 0.6179030 0.4486324 0.3559901 0.3242083 0.5824228 0.4633715 0.6044657
#> [2794] 0.6288954 0.4612112 0.6369108 0.6127666 0.5748209 0.4525067 0.6084384
#> [2801] 0.5090356 0.5356711 0.5143658 0.6038173 0.6003113 0.5193530 0.6225299
#> [2808] 0.4741065 0.3163947 0.4017793 0.5699601 0.3027305 0.6232452 0.4736269
#> [2815] 0.6234541 0.4203878 0.4407043 0.6353988 0.4625113 0.4163441 0.6138282
#> [2822] 0.5654725 0.5627061 0.4218931 0.6219728 0.4306435 0.4711454 0.6243603
#> [2829] 0.4397251 0.5071188 0.4242151 0.5399605 0.5569046 0.6369108 0.4907641
#> [2836] 0.4521336 0.5681173 0.5589134 0.4416643 0.6314017 0.6027248 0.6256455
#> [2843] 0.6065522 0.5951917 0.3970642 0.6144316 0.6323075 0.4597727 0.2996693
#> [2850] 0.5923476 0.3823868 0.4746157 0.5587625 0.4883671 0.6239689 0.5055273
#> [2857] 0.3233355 0.3289026 0.6348579 0.5175600 0.3097859 0.5682573 0.5171355
#> [2864] 0.5263611 0.4961333 0.6355278 0.4899605 0.5587661 0.6364836 0.6356744
#> [2871] 0.5896983 0.5628344 0.6115351 0.3166359 0.3925647 0.4944739 0.3410302
#> [2878] 0.3986425 0.5596449 0.6210364 0.4963722 0.5400796 0.5032981 0.4194727
#> [2885] 0.6317336 0.5158437 0.4630111 0.6327370 0.5136500 0.5448528 0.4352159
#> [2892] 0.4496192 0.4631308 0.6045498 0.3904310 0.2384648 0.5567499 0.3982300
#> [2899] 0.6115751 0.4132332 0.1923618 0.3716462 0.4327556 0.5582764 0.4044002
#> [2906] 0.5286098 0.3085107 0.5557091 0.4987535 0.6229158 0.4430536 0.3789448
#> [2913] 0.4079433 0.4071562 0.6342753 0.4998524 0.5685970 0.6255748 0.6202021
#> [2920] 0.6123657 0.6343138 0.2888582 0.6051335 0.5502521 0.5386897 0.3877998
#> [2927] 0.5855334 0.5042760 0.5173239 0.6208301 0.5148232 0.5538765 0.6354516
#> [2934] 0.6342215 0.3621144 0.6302531 0.6147636 0.6269297 0.4850646 0.6165237
#> [2941] 0.5225210 0.5655851 0.5068183 0.4243367 0.6308344 0.5344120 0.5213513
#> [2948] 0.4093291 0.6311828 0.5404245 0.2934971 0.5384165 0.4321563 0.1719303
#> [2955] 0.3438455 0.5928896 0.2673648 0.4275921 0.5135557 0.5962283 0.6096473
#> [2962] 0.6088715 0.3137898 0.6248784 0.5748364 0.6264696 0.5111422 0.4717987
#> [2969] 0.5564482 0.5070713 0.6103054 0.6241500 0.4048767 0.5560480 0.5395650
#> [2976] 0.5608116 0.4005732 0.4039873 0.3985610 0.5897662 0.5236547 0.5294643
#> [2983] 0.5555485 0.2686676 0.6129236 0.6303586 0.6229619 0.5079633 0.5588708
#> [2990] 0.2465271 0.4056581 0.5990539 0.6055871 0.5484728 0.5684715 0.3977737
#> [2997] 0.6052879 0.5870691 0.6369108 0.3714917 0.3617024 0.4506421 0.3886281
#> [3004] 0.6345645 0.5099151 0.6035899 0.5571339 0.4133292 0.6290400 0.3908397
#> [3011] 0.5199216 0.5669071 0.4821392 0.4557725 0.6298858 0.5607908 0.4383319
#> [3018] 0.4594815 0.6002382 0.5865120 0.6124672 0.6284240 0.5671486 0.6283997
#> [3025] 0.5420964 0.6312181 0.5443383 0.5475840 0.6096805 0.4621868 0.5127399
#> [3032] 0.5023746 0.6240625 0.4405421 0.6358151 0.5611632 0.6112966 0.6223396
#> [3039] 0.6348751 0.2195770 0.5095286 0.2744031 0.6279578 0.4791295 0.4960339
#> [3046] 0.5726481 0.4870062 0.6252854 0.6217184 0.3224919 0.5416352 0.5986085
#> [3053] 0.5081319 0.5394117 0.5940570 0.5224777 0.5965999 0.3017164 0.3742333
#> [3060] 0.5910918 0.5926263 0.3108269 0.4880132 0.3651566 0.2838822 0.2789641
#> [3067] 0.5634284 0.3472466 0.5993306 0.5556455 0.5686328 0.3035809 0.5957461
#> [3074] 0.5594415 0.3396792 0.5740969 0.4157521 0.6025734 0.5079468 0.4356411
#> [3081] 0.2594919 0.5167535 0.5379006 0.6290382 0.6168478 0.6193747 0.5276557
#> [3088] 0.6338931 0.3479293 0.5301660 0.6360772 0.5862152 0.5814192 0.6334302
#> [3095] 0.4357118 0.4601635 0.5095286 0.3715079 0.6036997 0.6328124 0.5964095
#> [3102] 0.5123208 0.6331428 0.5971448 0.5681532 0.3454623 0.4296368 0.5661229
#> [3109] 0.3335102 0.6303672 0.4151468 0.5677440 0.3544808 0.4501483 0.4639169
#> [3116] 0.6034620 0.6338234 0.5328672 0.6278692 0.5794097 0.5072945 0.4090430
#> [3123] 0.5719485 0.4080424 0.6179276 0.5691658 0.5099914 0.6072239 0.5008270
#> [3130] 0.4621762 0.5526877 0.3515771 0.5803423 0.4991440 0.6203655 0.6177685
#> [3137] 0.5983908 0.4976037 0.4783437 0.4416166 0.4589906 0.5377242 0.6225375
#> [3144] 0.5976709 0.5910400 0.5601841 0.4608316 0.3674938 0.4488235 0.3917213
#> [3151] 0.3104710 0.5340121 0.4837890 0.3847198 0.6233540 0.3821498 0.4922636
#> [3158] 0.5625742 0.3053865 0.2059272 0.6138008 0.6296140 0.4190927 0.4377943
#> [3165] 0.3663281 0.6322990 0.5808402 0.5505159 0.5230130 0.3766007 0.6251148
#> [3172] 0.5659124 0.6064486 0.6155257 0.6291093 0.4067116 0.5380887 0.4609470
#> [3179] 0.4597309 0.3747536 0.5837643 0.5841122 0.3302312 0.2725684 0.6314366
#> [3186] 0.6278466 0.5580728 0.5006872 0.3118346 0.5884025 0.5890881 0.2866696
#> [3193] 0.6034145 0.5482151 0.6100361 0.5190660 0.6165498 0.4509204 0.3848811
#> [3200] 0.6152106 0.6307014 0.2237130 0.5985014 0.5512994 0.5334150 0.1803444
#> [3207] 0.4588482 0.4145498 0.5435003 0.3790794 0.4278705 0.6363434 0.5348791
#> [3214] 0.4457033 0.4776948 0.5582345 0.6280068 0.6267010 0.6297780 0.3352116
#> [3221] 0.6368387 0.4434085 0.2929112 0.4487646 0.4572669 0.4861863 0.5308548
#> [3228] 0.5742667 0.6273032 0.6125879 0.6345752 0.5228253 0.3958805 0.6302969
#> [3235] 0.5024214 0.5680385 0.5133720 0.5730096 0.4488147 0.6264545 0.6304814
#> [3242] 0.5562239 0.4158633 0.5986553 0.3875800 0.3328418 0.5796920 0.4597942
#> [3249] 0.5539288 0.6257194 0.4707294 0.6495794 0.4429395 0.3819943 0.2692290
#> [3256] 0.4328840 0.5053410 0.5177133 0.5302314 0.3344451 0.6363021 0.2036768
#> [3263] 0.6334446 0.6324910 0.3900183 0.3815353 0.3945678 0.6171494 0.4596165
#> [3270] 0.6118362 0.4756412 0.3579675 0.6301998 0.6302866 0.4723117 0.5450691
#> [3277] 0.6357871 0.6333334 0.6022673 0.4981844 0.5934650 0.2035852 0.4804531
#> [3284] 0.6267643 0.6307549 0.4338189 0.6369108 0.5105342 0.5846115 0.6304737
#> [3291] 0.2557048 0.6146319 0.2950681 0.6338530 0.3790832 0.6118520 0.5215659
#> [3298] 0.6013391 0.4584548 0.4614097 0.4965612 0.5895393 0.4908038 0.5638529
#> [3305] 0.4570524 0.4718184 0.5656118 0.4917132 0.6016130 0.5255680 0.6090972
#> [3312] 0.6276300 0.3197590 0.4975688 0.5196689 0.5945205 0.5995142 0.4732162
#> [3319] 0.6347631 0.6154356 0.5883512 0.3593984 0.3603013 0.6108359 0.5565188
#> [3326] 0.3504557 0.4637157 0.5644467 0.3916366 0.5309414 0.5541206 0.2761515
#> [3333] 0.5050659 0.3887957 0.6103001 0.2496128 0.6210863 0.5507960 0.4603482
#> [3340] 0.4644549 0.5505461 0.5351022 0.4396567 0.6119026 0.5840387 0.3821883
#> [3347] 0.5532922 0.6324968 0.3412061 0.6174927 0.5581978 0.4561433 0.1900342
#> [3354] 0.6109202 0.2664588 0.5785975 0.4593908 0.4720373 0.5789952 0.5740628
#> [3361] 0.2501549 0.4661382 0.2340146 0.5129842 0.4430327 0.6102550 0.5338133
#> [3368] 0.5626097 0.4302414 0.5053798 0.5969124 0.6265972 0.5409163 0.4743201
#> [3375] 0.6274947 0.5572615 0.3997273 0.5376820 0.4708868 0.3462822 0.3702092
#> [3382] 0.5243804 0.4324928 0.6068251 0.5142481 0.5146388 0.3595212 0.5439350
#> [3389] 0.5547341 0.6242028 0.5527905 0.4887320 0.6238440 0.5088923 0.4925698
#> [3396] 0.5258618 0.2644215 0.2959665 0.2390594 0.6214894 0.5131401 0.4367473
#> [3403] 0.5594035 0.4392866 0.3289067 0.4991782 0.5900273 0.5582313 0.5958696
#> [3410] 0.1783524 0.4638381 0.6085897 0.3491629 0.6369108 0.4482922 0.5533515
#> [3417] 0.6270800 0.6183351 0.4603824 0.6144154 0.5482003 0.4882809 0.5286799
#> [3424] 0.5688092 0.5257237 0.6312886 0.4028365 0.6307420 0.5058929 0.6369108
#> [3431] 0.4867321 0.6280841 0.4638766 0.4692416 0.4680602 0.4067302 0.6217435
#> [3438] 0.4947219 0.5732658 0.2212103 0.5292180 0.5768820 0.6142669 0.4843779
#> [3445] 0.6014936 0.6282565 0.5870398 0.5176406 0.2780195 0.5906364 0.5911303
#> [3452] 0.5119556 0.4438669 0.3201005 0.6360718 0.5878897 0.5072843 0.4329308
#> [3459] 0.6320790 0.3777064 0.6232557 0.6113991 0.6357421 0.5396174 0.5044311
#> [3466] 0.5035844 0.4840895 0.6353533 0.2919898 0.4033768 0.5203890 0.5718186
#> [3473] 0.5580551 0.6255799 0.6298429 0.4672682 0.4173980 0.5744087 0.3554071
#> [3480] 0.3801903 0.6009989 0.4203370 0.5689736 0.4333711 0.6134537 0.5824611
#> [3487] 0.6101275 0.6186815 0.5184270 0.4890491 0.4986741 0.6306462 0.6258279
#> [3494] 0.1934888 0.2711501 0.6087001 0.6143710 0.4514379 0.3784509 0.3706566
#> [3501] 0.3196433 0.4195022 0.5921589 0.5845496 0.4913532 0.6142650 0.4766354
#> [3508] 0.4109640 0.5154901 0.4209179 0.5454195 0.3837028 0.6210052 0.4722502
#> [3515] 0.6259797 0.6149687 0.5235773 0.5963871 0.5787434 0.3295699 0.6029244
#> [3522] 0.3187020 0.6157323 0.5978473 0.4609689 0.2972402 0.5722205 0.6081278
#> [3529] 0.5779520 0.6329913 0.5633118 0.6232526 0.4458375 0.5635676 0.6115756
#> [3536] 0.4872313 0.4030873 0.3909856 0.3636744 0.6360061 0.6187448 0.6059969
#> [3543] 0.6293557 0.6075114 0.5300764 0.3177625 0.6083302 0.3415714 0.6258764
#> [3550] 0.3821465 0.5211318 0.5429675 0.6245179 0.6100449 0.4443913 0.6001376
#> [3557] 0.5687253 0.4239923 0.5306717 0.5289824 0.5131619 0.5348001 0.4716854
#> [3564] 0.5273152 0.6190847 0.3692699 0.4158002 0.6147543 0.5709569 0.4924393
#> [3571] 0.4738942 0.3931246 0.6262769 0.4071583 0.3523651 0.5916646 0.5600959
#> [3578] 0.5039552 0.6101996 0.5691569 0.6267863 0.5346641 0.2540281 0.5807615
#> [3585] 0.6263506 0.6249913 0.4986141 0.3438651 0.6186398 0.4323231 0.4551300
#> [3592] 0.3688903 0.5478052 0.6247933 0.3984849 0.3794920 0.6281258 0.5527840
#> [3599] 0.5204508 0.6369108 0.2888142 0.5180760 0.5982481 0.5068121 0.5512994
#> [3606] 0.5676698 0.5173553 0.4727649 0.4305220 0.5690876 0.5739832 0.5130374
#> [3613] 0.6093082 0.5640941 0.5193017 0.5142177 0.4586762 0.4122247 0.1951309
#> [3620] 0.5673817 0.6024858 0.6076703 0.6297657 0.6247522 0.3913189 0.4630798
#> [3627] 0.3647692 0.6274486 0.5203271 0.4477269 0.3184937 0.5659717 0.6316617
#> [3634] 0.4527461 0.5012234 0.5209882 0.3635174 0.4124999 0.5881356 0.3273298
#> [3641] 0.6145011 0.5025678 0.5238868 0.6008314 0.4254784 0.3009745 0.4344744
#> [3648] 0.3203702 0.6105184 0.2781995 0.4907880 0.5371474 0.4828770 0.4141057
#> [3655] 0.3591675 0.4600314 0.5740443 0.6317732 0.6264696 0.3522694 0.5435439
#> [3662] 0.5189348 0.6349707 0.5190596 0.6269495 0.4769063 0.5008455 0.6024160
#> [3669] 0.6292065 0.5598543 0.6025717 0.6185854 0.5750400 0.4402955 0.6183107
#> [3676] 0.3596943 0.4837737 0.6278886 0.6338059 0.5923131 0.6268714 0.6263033
#> [3683] 0.4567942 0.6359032 0.6150229 0.3624116 0.5677802 0.5826784 0.5461835
#> [3690] 0.2898670 0.4027546 0.3015389 0.5658280 0.3826523 0.3229918 0.6184371
#> [3697] 0.3390099 0.6179914 0.5653372 0.5120228 0.3063483 0.4882003 0.6114407
#> [3704] 0.4998795 0.5148232 0.4683374 0.5273384 0.5112148 0.5181646 0.6350428
#> [3711] 0.5710857 0.6308946 0.5900231 0.5454900 0.4417450 0.6369108 0.4792826
#> [3718] 0.6361795 0.6182975 0.2573387 0.5486132 0.6298325 0.6305656 0.4891773
#> [3725] 0.4490358 0.6334590 0.3764663 0.5461177 0.3771639 0.4458398 0.5103506
#> [3732] 0.5418204 0.3940084 0.5972536 0.6135494 0.6110589 0.4672152 0.6369108
#> [3739] 0.2304307 0.3838442 0.4786771 0.4425761 0.4701300 0.5338457 0.6160967
#> [3746] 0.5732197 0.5323325 0.4362384 0.6291739 0.6339472 0.6257011 0.2807371
#> [3753] 0.6212658 0.6359424 0.3601383 0.2391376 0.5986598 0.4068304 0.5065446
#> [3760] 0.5262066 0.5464248 0.6348022 0.5064300 0.6104051 0.6093448 0.6136194
#> [3767] 0.4458391 0.5466999 0.6332655 0.4860483 0.5001804 0.4086992 0.5221530
#> [3774] 0.5055407 0.2713342 0.4620180 0.5919785 0.4765552 0.4461742 0.5328087
#> [3781] 0.5046376 0.3940997 0.5412287 0.4774405 0.6027338 0.3607021 0.6151029
#> [3788] 0.2887006 0.3381159 0.6151656 0.4549825 0.6363693 0.4510404 0.6257585
#> [3795] 0.6227046 0.5219277 0.6360445 0.6246936 0.3346736 0.3931921 0.6121331
#> [3802] 0.5879669 0.4815792 0.6269359 0.5744716 0.5015418 0.6219145 0.6074028
#> [3809] 0.5458406 0.6042864 0.1949182 0.6344200 0.4285268 0.4021496 0.5459985
#> [3816] 0.5440072 0.5167992 0.5725040 0.6346673 0.4528059 0.5354918 0.6313302
#> [3823] 0.5533901 0.6320347 0.5346770 0.3778646 0.3983798 0.4053741 0.5454894
#> [3830] 0.3564750 0.6317244 0.5235995 0.4143179 0.6146244 0.6330921 0.5218607
#> [3837] 0.4810959 0.4951411 0.3472466 0.5714983 0.4426742 0.4485936 0.5669160
#> [3844] 0.3379694 0.4388097 0.6265950 0.5072963 0.6341885 0.5310448 0.4546521
#> [3851] 0.5670797 0.5880017 0.4453727 0.3691828 0.3198036 0.4542414 0.4507046
#> [3858] 0.4120320 0.5230748 0.5833598 0.4710414 0.5990222 0.6182455 0.4936834
#> [3865] 0.4132812 0.3188285 0.5606089 0.3754223 0.3316383 0.2239015 0.4997745
#> [3872] 0.6039269 0.3664721 0.3172271 0.3389915 0.5709795 0.5308525 0.6143161
#> [3879] 0.4797093 0.5480390 0.6173362 0.5807127 0.4503719 0.5936095 0.4976489
#> [3886] 0.6036776 0.4701906 0.5367736 0.5135650 0.5423534 0.3282334 0.6177946
#> [3893] 0.4372134 0.6291095 0.4599808 0.5885722 0.4640143 0.6074935 0.4386897
#> [3900] 0.6097505 0.4728901 0.4274534 0.5730902 0.5804482 0.6082069 0.3463848
#> [3907] 0.2280310 0.5159451 0.5207173 0.3562619 0.4189677 0.5623567 0.4752830
#> [3914] 0.5372960 0.4442137 0.1096242 0.5952248 0.3999830 0.4708113 0.4906761
#> [3921] 0.5241309 0.4235111 0.3216143 0.5868983 0.2373504 0.6036982 0.5214865
#> [3928] 0.6338369 0.6295644 0.5951058 0.5980154 0.4561858 0.4785608 0.4656967
#> [3935] 0.4286087 0.5705288 0.4513460 0.3284322 0.5732265 0.5934651 0.5393472
#> [3942] 0.6333447 0.5270629 0.5811507 0.6368466 0.6121560 0.5926069 0.5306947
#> [3949] 0.5022639 0.6348286 0.6255398 0.5106745 0.6355249 0.5975028 0.5681861
#> [3956] 0.4883255 0.6029597 0.6337886 0.5271921 0.3306314 0.5387905 0.6367995
#> [3963] 0.5721803 0.5902596 0.4411248 0.5619801 0.5831641 0.2448907 0.3805974
#> [3970] 0.3331106 0.4781762 0.6091541 0.3075213 0.5210538 0.4882982 0.4362384
#> [3977] 0.3286331 0.4117137 0.6365423 0.5811542 0.3618725 0.5165962 0.6155563
#> [3984] 0.5615394 0.6096728 0.6213189 0.3911545 0.5708004 0.6216696 0.4205463
#> [3991] 0.4111181 0.4160105 0.3819558 0.6206483 0.6264399 0.5383152 0.3897399
#> [3998] 0.4991720 0.5197970 0.5634288 0.5210522 0.5040386 0.6293934 0.5012810
#> [4005] 0.4871132 0.5995844 0.5744843 0.4879010 0.5109103 0.6214416 0.5885462
#> [4012] 0.6364187 0.3569088 0.3303843 0.6298272 0.5987585 0.4595496 0.4577403
#> [4019] 0.5375421 0.6022652 0.5471613 0.6157552 0.5631221 0.5489575 0.5854844
#> [4026] 0.6146840 0.5993340 0.5354358 0.6313247 0.4850183 0.6092612 0.3824841
#> [4033] 0.4153150 0.6301096 0.4463966 0.6368905 0.5619801 0.5929990 0.1891379
#> [4040] 0.6159420 0.5882561 0.4971580 0.6272482 0.4797451 0.4364979 0.2852418
#> [4047] 0.6112168 0.6097117 0.3426063 0.3940487 0.3671545 0.2281678 0.5540274
#> [4054] 0.5569936 0.6210609 0.5320492 0.6186181 0.5088422 0.3541912 0.5356374
#> [4061] 0.4103359 0.6194024 0.5991017 0.3088514 0.5364506 0.4909016 0.6007847
#> [4068] 0.4947505 0.6000907 0.6016564 0.3720278 0.5319103 0.3095870 0.3828208
#> [4075] 0.3868484 0.6253719 0.6213590 0.3877951 0.5769796 0.1563225 0.3823037
#> [4082] 0.2574433 0.2584788 0.5559602 0.6310600 0.6117479 0.4054532 0.6281007
#> [4089] 0.3217017 0.5700504 0.6095706 0.5057166 0.6362074 0.5821485 0.5351063
#> [4096] 0.5265291 0.6242225 0.4471759 0.3239183 0.6083607 0.6331848 0.4133362
#> [4103] 0.5790280 0.5542555 0.2706861 0.3517854 0.4602140 0.5858597 0.4963301
#> [4110] 0.5350698 0.5185547 0.5362842 0.6105639 0.5506648 0.6070847 0.5905015
#> [4117] 0.3400886 0.5127968 0.6184910 0.6352646 0.6214445 0.5350411 0.4190502
#> [4124] 0.5100469 0.4192089 0.5210352 0.6254580 0.3626406 0.4579500 0.5548120
#> [4131] 0.6245710 0.6369108 0.4491597 0.6152400 0.5827682 0.4154780 0.6313743
#> [4138] 0.6033332 0.5402670 0.4852741 0.5821674 0.4409336 0.5330260 0.3593879
#> [4145] 0.6325162 0.3559698 0.5669815 0.3657825 0.6315017 0.6033297 0.6289151
#> [4152] 0.4489786 0.3434386 0.5297717 0.5211334 0.4586671 0.5610721 0.5386226
#> [4159] 0.6146294 0.5708804 0.2067429 0.3633704 0.6358210 0.4669696 0.6149928
#> [4166] 0.4484756 0.3863297 0.3619722 0.6369108 0.6221488 0.6020897 0.3358729
#> [4173] 0.5948720 0.5798500 0.2406030 0.3974332 0.6079440 0.5226196 0.5688166
#> [4180] 0.5905532 0.2499876 0.4897735 0.6314145 0.5682291 0.5005882 0.5898911
#> [4187] 0.5852743 0.6142349 0.5592699 0.5126985 0.4888525 0.5631480 0.3253443
#> [4194] 0.5005875 0.5238485 0.5886657 0.1757015 0.5821666 0.6115085 0.2789168
#> [4201] 0.6234452 0.5963999 0.5315795 0.3556427 0.6349984 0.6231758 0.4628755
#> [4208] 0.5352994 0.4345674 0.6173513 0.3958364 0.5437832 0.5318995 0.6327886
#> [4215] 0.3799331 0.3649243 0.3455734 0.5361295 0.6025533 0.5849084 0.6233075
#> [4222] 0.5879176 0.3787164 0.5570877 0.3864206 0.5231223 0.4420120 0.6089302
#> [4229] 0.5915629 0.4611330 0.5461325 0.3948583 0.6271071 0.6299050 0.4518072
#> [4236] 0.5324882 0.3593799 0.6225884 0.4667406 0.5125029 0.6116716 0.6114343
#> [4243] 0.6044324 0.3782661 0.5953793 0.6345289 0.5888892 0.6366933 0.5159320
#> [4250] 0.6150807 0.4696735 0.4777473 0.5189876 0.4536015 0.4584951 0.5126899
#> [4257] 0.5193450 0.3611183 0.6091396 0.3472565 0.3719584 0.4319100 0.6357094
#> [4264] 0.6333034 0.4653774 0.4400602 0.4621186 0.4426158 0.5661902 0.5839838
#> [4271] 0.2645765 0.6324547 0.5594487 0.5191581 0.5995420 0.6167252 0.6301371
#> [4278] 0.3524658 0.4655258 0.6077880 0.6091168 0.6223694 0.5859279 0.4272667
#> [4285] 0.4019224 0.3721494 0.3183092 0.5095286 0.5667509 0.5270509 0.5940197
#> [4292] 0.3186358 0.4287825 0.5165985 0.5190419 0.4599610 0.5566829 0.4905410
#> [4299] 0.3971027 0.4803544 0.6310357 0.5494916 0.6298559 0.2492812 0.5510201
#> [4306] 0.4532847 0.5763572 0.5491684 0.6313902 0.5619801 0.3931553 0.5650550
#> [4313] 0.3122940 0.5692970 0.5984226 0.5319232 0.4091926 0.3653559 0.4828112
#> [4320] 0.6251857 0.5925446 0.4778104 0.4343598 0.3202169 0.4033377 0.5844957
#> [4327] 0.6365424 0.5905603 0.3286735 0.5488332 0.4756731 0.6112871 0.5458522
#> [4334] 0.5400716 0.5455382 0.5113755 0.6216149 0.4718332 0.5960860 0.5579662
#> [4341] 0.5280773 0.5785953 0.5354530 0.5313244 0.6315676 0.4114969 0.5460253
#> [4348] 0.3384057 0.3755426 0.6293489 0.4234696 0.5346214 0.4891615 0.5293832
#> [4355] 0.3134098 0.3298351 0.4334262 0.3053160 0.4333683 0.4498248 0.2942888
#> [4362] 0.5267424 0.4718064 0.3426197 0.6175683 0.4980749 0.6028394 0.4555449
#> [4369] 0.5794514 0.6230575 0.6140910 0.5940677 0.6164819 0.5969295 0.4840193
#> [4376] 0.5507044 0.4063337 0.5953641 0.2969890 0.6296134 0.6339094 0.6316092
#> [4383] 0.4610858 0.5707877 0.5960529 0.5532033 0.5864437 0.5058503 0.5067191
#> [4390] 0.4656489 0.3934297 0.4464433 0.6126904 0.5586972 0.4549783 0.5205905
#> [4397] 0.5917039 0.5376031 0.6115120 0.6137478 0.4728771 0.6367012 0.6354974
#> [4404] 0.5907458 0.5212004 0.1453867 0.3099957 0.4632648 0.6153667 0.4882820
#> [4411] 0.5779257 0.3791397 0.3276900 0.2058801 0.6148393 0.5804747 0.5633590
#> [4418] 0.6062132 0.5431335 0.5378971 0.5741995 0.4739579 0.5654599 0.3985539
#> [4425] 0.3648369 0.5797374 0.6302223 0.5637266 0.5317499 0.5631197 0.6249311
#> [4432] 0.3670436 0.5421666 0.5177148 0.6150288 0.5506806 0.6158587 0.6363558
#> [4439] 0.5011017 0.6174243 0.6170207 0.3096613 0.4519753 0.6253340 0.5339557
#> [4446] 0.6227530 0.3997533 0.4141252 0.6303103 0.6356037 0.5432043 0.3171340
#> [4453] 0.4081392 0.4183592 0.6346882 0.6242868 0.6357511 0.4238851 0.4617959
#> [4460] 0.4594034 0.4140469 0.1655853 0.3882195 0.4241209 0.4714461 0.5742079
#> [4467] 0.4309716 0.5563651 0.5703034 0.3659860 0.5588221 0.4385502 0.5708886
#> [4474] 0.6321929 0.5283157 0.6262956 0.6319170 0.4950129 0.3416202 0.5309662
#> [4481] 0.5743679 0.6292665 0.4299349 0.4132153 0.5252467 0.6360264 0.5352778
#> [4488] 0.4776233 0.3584116 0.6347966 0.4082162 0.6320744 0.4619120 0.4345062
#> [4495] 0.3689240 0.3254399 0.2629398 0.6329417 0.6178316 0.6355889 0.4694998
#> [4502] 0.3577770 0.4467329 0.6316797 0.4552550 0.3014941 0.3506316 0.5509615
#> [4509] 0.6165752 0.6155828 0.3952426 0.6231410 0.4941222 0.4262297 0.5909184
#> [4516] 0.6275149 0.3327607 0.2413231 0.5084013 0.5487231 0.3510117 0.5847262
#> [4523] 0.3109435 0.6236039 0.5061210 0.6203270 0.4380845 0.5102533 0.4588600
#> [4530] 0.4352793 0.5615086 0.5676221 0.5649152 0.5573477 0.6164403 0.6157229
#> [4537] 0.4667250 0.1955739 0.5856593 0.6099105 0.6133302 0.5434856 0.5968403
#> [4544] 0.2203402 0.4824642 0.6181228 0.2903717 0.4198611 0.4820499 0.5962296
#> [4551] 0.4932244 0.2393185 0.4020831 0.4982671 0.6057857 0.5034477 0.6279750
#> [4558] 0.2713626 0.5424772 0.6195946 0.6105123 0.3590346 0.5005369 0.5442367
#> [4565] 0.6204364 0.6278044 0.5157551 0.5195189 0.4862176 0.5613420 0.6348138
#> [4572] 0.5259156 0.6052681 0.6311521 0.4202718 0.4789861 0.4021171 0.4907478
#> [4579] 0.5068070 0.4774208 0.3757992 0.6309771 0.4676610 0.6339781 0.6156575
#> [4586] 0.6313287 0.5741789 0.5496487 0.5668984 0.5787884 0.6360288 0.4448742
#> [4593] 0.5080986 0.4768572 0.3220227 0.4328497 0.5398280 0.4244482 0.4984419
#> [4600] 0.6403120 0.4432513 0.6304976 0.6029339 0.6167268 0.4375318 0.3654559
#> [4607] 0.4385041 0.5980249 0.3912158 0.5945873 0.4852029 0.3245049 0.4874067
#> [4614] 0.4018721 0.5553939 0.5500922 0.6129651 0.4642401 0.5101789 0.4742024
#> [4621] 0.6219375 0.6026106 0.6321751 0.4695805 0.4160229 0.5568305 0.6113273
#> [4628] 0.6275907 0.3325223 0.4086579 0.5625746 0.5788109 0.4819510 0.6060632
#> [4635] 0.6131685 0.6287674 0.5920907 0.5902788 0.3348104 0.5930339 0.4435126
#> [4642] 0.6325742 0.5520964 0.3211749 0.1983036 0.5341871 0.2085808 0.4727419
#> [4649] 0.3336639 0.3900748 0.5214936 0.5706708 0.5013277 0.3858837 0.4766791
#> [4656] 0.6126389 0.4846970 0.5788288 0.5629777 0.5684516 0.6294651 0.4517216
#> [4663] 0.5416330 0.6209924 0.6234926 0.4370413 0.6141639 0.6227819 0.6302019
#> [4670] 0.6323716 0.4290761 0.6347439 0.6317379 0.5784941 0.6336007 0.6191450
#> [4677] 0.6293328 0.5522995 0.6335431 0.5039746 0.5731650 0.3763975 0.4044631
#> [4684] 0.5580467 0.5857001 0.6026156 0.6234937 0.4355525 0.5003681 0.5378488
#> [4691] 0.5577191 0.6307674 0.5808849 0.3715707 0.5496904 0.3419281 0.3590336
#> [4698] 0.5013951 0.5741605 0.5343952 0.6021087 0.6279215 0.6115896 0.4816847
#> [4705] 0.4895299 0.6288869 0.4722082 0.4253645 0.6293551 0.4861137 0.5725140
#> [4712] 0.6128910 0.5499568 0.5766699 0.5737478 0.6013910 0.6201383 0.4055715
#> [4719] 0.6103766 0.3503273 0.5079903 0.6256650 0.5440042 0.5223941 0.6327002
#> [4726] 0.6297473 0.5307324 0.5378901 0.4627010 0.4004261 0.6346731 0.5944675
#> [4733] 0.5505734 0.3934000 0.6132118 0.4062199 0.3738838 0.2694507 0.2024100
#> [4740] 0.3773575 0.3970910 0.5674371 0.5168237 0.5356881 0.5632945 0.2549956
#> [4747] 0.5893062 0.4354320 0.5173753 0.5339524 0.3419733 0.4814350 0.5846619
#> [4754] 0.4827982 0.3627688 0.6336900 0.6265523 0.4631220 0.5312449 0.5600758
#> [4761] 0.5098196 0.6259502 0.2941946 0.4874665 0.4620310 0.4456736 0.6235098
#> [4768] 0.5985788 0.3608285 0.4104361 0.5682696 0.5566829 0.5499325 0.6050597
#> [4775] 0.6291886 0.5017432 0.4141898 0.4269331 0.2958799 0.3414141 0.5880652
#> [4782] 0.6352991 0.5762759 0.3546547 0.5179477 0.3122735 0.4354973 0.2533707
#> [4789] 0.5122749 0.5035658 0.5989609 0.4978968 0.4229809 0.6090163 0.5080258
#> [4796] 0.6292714 0.6023581 0.5267276 0.3833752 0.5481162 0.4867612 0.4449094
#> [4803] 0.4936850 0.4781572 0.5758534 0.5427583 0.5534021 0.6247452 0.3017952
#> [4810] 0.5540699 0.6078572 0.5394590 0.4751269 0.4510127 0.3561678 0.6199945
#> [4817] 0.6353877 0.5369594 0.3569627 0.5326532 0.5731053 0.2659282 0.5882814
#> [4824] 0.3637889 0.4868934 0.4664551 0.3748456 0.6262752 0.5601564 0.4942087
#> [4831] 0.6097255 0.4618200 0.5521199 0.5914820 0.4951329 0.5597413 0.4215587
#> [4838] 0.3279097 0.4866200 0.5289026

# how the calculation is done - first observation
fs_area(buildings[1,]) / fs_area(fs_mbc(buildings[1,]))

#> 0.5732678 [1]