List footprint summary metrics

Helper functions to list footprint metric names and units

get_fs_metrics(short_names, group = NULL)

list_fs(col_names = NULL)

get_fs_units(metrics = NULL)

get_fs_group(metrics = NULL)

Arguments

short_names
Character vector of short names to lookup fs_ function names.

group
Character vector to return a group of metrics (e.g. "area").

col_names
Character vector to select columns from the table of metrics.

metrics
Character vector of fs_ function names to look up the default units.

Value

Vector or data.frame with selected footprint metric function names and/or units.

Details

These convenience function access the internal data in foot::fs_footprint_metrics. They provide easy look-up access for the function names, the default units of measurement, or to return columns. Arguments are optional and if omitted the full set of metrics will be returned.