Kampala building footprint sample

Sample of building footprint polygons from a small region of Kampala, Uganda with 4840 structures represented. In addition the dataset contains a 3 arc-second grid to serve as a template for gridding values and polygons of "administrative boundaries" and "survey clusters."

The building data were produced by Microsoft Bing Maps. The mastergrid is aligned to gridded data produced by WorldPop. The adminZones and cluster datasets are purely artificial and for demonstration purposes only.

data(kampala)

Format

A list with four objects.

"buildings" - polygons of building footprints in sf format

"mastergrid" - geoTiff RasterLayer

"adminZones" - polygons in sf format for zonal statistics

"clusters" - 10 small polygons in sf format for sample sites

Source

Uganda-Tanzania Building Footprints