List footprint summary metrics

Helper function to list footprint characteristic names and built-in summary functions.

list_fs(what = "all", how = "all")

Arguments

what
Character vector of names of characteristics to look up (e.g. "area"). Alternatively, list 'all' or all characteristics except nearest neighbour distances ('nodist').

how
Character vector of summary functions to look up (e.g. "mean") or 'all' available metrics.

Value

Vector or data.frame with selected footprint metric function names and/or units.

Details

Provides an easy look-up for the built-in function names, and basic geometric characteristics used by foot. Supplying a "what" characteristics finds all built-in functions. Conversely, supplying a "how" function name returns are characteristics for which that summary is available. Arguments are optional and if both are omitted the full set of metrics will be returned.

Examples

# get the full list of all available 
list_fs()

#>         cols    funs
#> 1      angle      cv
#> 2      angle entropy
#> 3      angle     max
#> 4      angle    mean
#> 5      angle  median
#> 6      angle     min
#> 7      angle      sd
#> 8       area      cv
#> 9       area     max
#> 10      area    mean
#> 11      area  median
#> 12      area     min
#> 13      area      sd
#> 14      area     sum
#> 15   compact      cv
#> 16   compact     max
#> 17   compact    mean
#> 18   compact  median
#> 19   compact     min
#> 20   compact      sd
#> 21    nndist      cv
#> 22    nndist     max
#> 23    nndist    mean
#> 24    nndist  median
#> 25    nndist     min
#> 26    nndist nnindex
#> 27    nndist      sd
#> 28 perimeter      cv
#> 29 perimeter     max
#> 30 perimeter    mean
#> 31 perimeter  median
#> 32 perimeter     min
#> 33 perimeter      sd
#> 34 perimeter     sum
#> 35   settled  binary
#> 36   settled   count
#> 37     shape      cv
#> 38     shape     max
#> 39     shape    mean
#> 40     shape  median
#> 41     shape     min
#> 42     shape      sd

# get all summary functions for "area"
list_fs(what="area")

#>   cols   funs
#> 1 area     cv
#> 2 area    max
#> 3 area   mean
#> 4 area median
#> 5 area    min
#> 6 area     sd
#> 7 area    sum

# get all characteristics relevant for entropy
list_fs(how="entropy")

#>    cols    funs
#> 1 angle entropy