Writing gridded data to binary files

Write data to binary files and structure as native raster files.

make_templateGrid(
  filename = NULL,
  datatype = "FLT8S",
  template = NULL,
  interleave = "BSQ",
  overwrite = FALSE
)

Arguments

filename
character string path to desired output file.

template
A filepath or a gridded data object to be used as a template to define the output file's resolution and positioning.

interleave
character string for band interleaving. Currently only 'BSQ' supported.

overwrite
logical. Should the output file be overwritten if it already exists? Default is False.

dataType
character string of raster data types. Default is FLT8S

Details

These functions provide a lightweight interface to binary writing demonstrated in the create_blank_raster package. The advantage of using binary files is the opportunity to using memory mapping (mmap) to support parallel writing.