Fast writing of binary files

write_imageBinary uses memory mapping functions to write gridded data to a binary output format. Binary outputs can be read natively but some software, or with the addition of header files, be converted to formats easily read by most software.

write_imageBinary(data, cellNumbers = 1:length(data), filename, mapMode)

Arguments

data
vector of

cellNumbers
vector of cell indices to update with data values

filename
character string of path to the output file

mapMode
character string of the raster data type.

Details

The function uses mmap create and update a binary output file on the hard drive.

The mapMode converts the raster data type strings to a CType used by mmap.

"LOG1S" mmap::logi8()

"INT1S" mmap::int8()

"INT1U" mmap::uint8()

"INT2S" mmap::int16()

"INT2U" mmap::uint16()

"INT4S" mmap::int32()

"INT4U" NA

"INT4S" mmap::real32()

"INT8S" mmap::real64()