Nearest neighbour distance calculation

Helper function to provide a distance calculation between spatial objects. The distance to the first nearest neighbour found (optionally within a maximum search radius) is returned.

fs_nndist(X, Y, maxSearch = 100, method = "poly", unit = "m")

Arguments

X
Spatial object of sf type, typically polygons or points.

Y
(Optional) Spatial object to measure distances to.

maxSearch
Maximum radius around X to search. Distance in meters. Default is 100. To ignore the search limit, set to `NULL`.

method
Either 'poly' or 'centroid' to assign a geometry method. See details. Default is 'poly'.

unit
Character abbreviation for the units to return from the distance calculation.

Details

If Y is omitted the nearest neighbour distances are found within X. Otherwise, the distance for each object in X to its nearest neighbour in Y is returned.

Use method to adjust which geometry of X and Y is used for distance calculations. When method='poly' distances are measured between polygon edges. When method='centroid', the centroids of building footprints are used instead. Centroid-based distance calculations are faster.

Providing a maximum search radius is strongly advised to speed up the calculation.

Examples

data("kampala", package="foot")

# get sample of buildings
buildings <- kampala$buildings
buildings <- buildings[sample(1:nrow(buildings), size=100, replace=F),]
clusters <- kampala$clusters

# calculate distance between buildings in m
fs_nndist(buildings, unit="m")

#> Units: [m]
#>   [1] 95.156073 65.601817 93.609789        NA 62.488049        NA        NA
#>   [8] 59.100124 46.189843        NA        NA 61.047564        NA 99.874799
#>  [15] 40.838343        NA 49.031866 47.966077 81.864406        NA        NA
#>  [22] 72.769564 72.769564        NA 40.838343 47.966077        NA 84.452095
#>  [29] 95.156073 51.058873 81.407425        NA 26.408178        NA        NA
#>  [36] 26.408178  6.255871 28.996263        NA        NA 74.070750 56.812414
#>  [43]        NA 33.712512        NA        NA 28.996263        NA        NA
#>  [50] 65.601817        NA 48.945229 51.058873 56.335511 49.031866 75.466904
#>  [57]        NA 33.712512 81.407425        NA        NA        NA        NA
#>  [64]        NA 32.857832        NA 92.682442 54.565714 59.100124 78.789825
#>  [71]        NA        NA 56.812414        NA        NA        NA        NA
#>  [78]        NA        NA 75.466904 61.047564        NA        NA        NA
#>  [85]        NA 68.010798  6.255871 32.857832 70.251191        NA 68.010798
#>  [92] 13.171564 84.452095 46.189843 13.171564 48.945229 81.864406 93.609789
#>  [99]        NA        NA

# calculate unrestricted distance 
# between buildings and another set of points
fs_nndist(buildings, sf::st_centroid(clusters), maxSearch=NULL)

#> Warning: st_centroid assumes attributes are constant over geometries of x

#> Warning: st_centroid does not give correct centroids for longitude/latitude data

#> Units: [m]
#>   [1]  565.3182  575.2489  504.6872  588.2953  347.4469  686.9973  913.9714
#>   [8]  648.2031  559.8968  625.5754  440.8780  715.5617  541.6824  737.0240
#>  [15]  493.5160  444.8671  496.0283  720.9820  983.2503  917.4102  732.0587
#>  [22]  580.3403  621.2782  740.2187  544.3737  756.0986  522.6995  490.5673
#>  [29]  702.6093  924.0940  647.9873  702.6034  629.1258  621.1267  599.2047
#>  [36]  671.9687  707.0049  629.2830  745.5526  679.7994  701.4090  327.6876
#>  [43]  813.2120  691.1722  667.1933  765.7186  660.6861  700.4227  636.3701
#>  [50]  607.0421  626.7668  342.6762  902.3689  913.9209  463.9412  584.4413
#>  [57]  708.2913  672.0993  748.4964  707.0380  340.0872 1075.9869  350.3166
#>  [64]  659.4719  454.7672  599.6658  406.7191  415.9892  651.2843  579.1138
#>  [71]  641.4879  659.4152  376.9941  483.4831  767.8711  642.5196  590.6956
#>  [78]  760.1477  582.7286  574.1047  751.5171  908.2384  791.5463  763.0176
#>  [85]  692.2449  598.3155  699.5359  405.8884  670.7742  530.1799  712.6713
#>  [92]  592.3538  526.0394  600.8611  610.7973  407.7618 1002.0050  558.2650
#>  [99]  501.0853  518.0380

# use footprint centroids
fs_nndist(buildings, method='centroid', unit='m')

#> Units: [m]
#>   [1]       NA 73.36415       NA       NA 83.76082       NA       NA 66.17950
#>   [9] 61.77284       NA       NA 71.77209       NA       NA 53.63681       NA
#>  [17] 63.03600 68.78486 91.96353       NA       NA 83.19872 83.19872       NA
#>  [25] 53.63681 68.78486       NA 94.12350       NA 65.76860       NA       NA
#>  [33] 47.35636       NA       NA 47.35636 37.43966 40.41447       NA       NA
#>  [41] 98.95258 65.42620       NA 52.52931       NA       NA 40.41447       NA
#>  [49]       NA 73.36415       NA 72.91177 65.76860 74.43525 63.03600 91.90861
#>  [57]       NA 52.52931       NA       NA       NA       NA       NA       NA
#>  [65] 45.11204       NA 99.85517 73.54571 66.17950 89.78993       NA       NA
#>  [73] 65.42620       NA       NA       NA       NA       NA       NA 91.90861
#>  [81] 71.77209       NA       NA       NA       NA       NA 37.43966 45.11204
#>  [89] 90.28790       NA       NA 27.67890 94.12350 61.77284 27.67890 72.91177
#>  [97] 91.96353       NA       NA       NA