Compactness index

Calculates an approximate measure of shape "compactness".

Calculates a measure of shape or "roundness".

fs_compact(X)

fs_shape(X, unit = NULL)

Arguments

X
polygons of building footprints of type sf.

unit
string indicating unit of measure. Passed to units::set_units.

Source

Polsby, Daniel D., and Robert D. Popper. 1991. “The Third Criterion: Compactness as a procedural safeguard against partisan gerrymandering.” Yale Law & Policy Review 9 (2): 301–353.

Value

numeric vector of compactness values ranging from 0 to 1 for each item in X.

numeric vector with shape index values ranging from 0 to 1 or each item in X.

Details

The compactness measure is the Polsby-Popper test based on the perimeter and area of each footprint.

$$ PP_i = \frac{4$\pi$ * A_z}{P_i^2} $$, where P and A are the perimeter and area of footprint 'i', respectively. Values closer to 1 indicate more compact shapes.

The shape index is calculated as the ratio of footprint polygon's area to the area of minimum bounding circle. The mbc is calculated using fs_mbc and lwgeom. Values closer 1 suggest more rounded shapes.

The function first looks for pre-calculated values of area in a field called fs_area. If not present, the areas are calculated in m^2.

Examples

data("kampala", package="foot")
buildings = kampala$buildings

fs_shape(buildings)

#>    [1] 0.5732710 0.3936445 0.5940952 0.4757789 0.5279643 0.4594503 0.5520442
#>    [8] 0.6186284 0.6146991 0.3191903 0.5831908 0.3734735 0.6199261 0.6284222
#>   [15] 0.5296343 0.5650630 0.6277424 0.3951911 0.4613214 0.4603272 0.5462177
#>   [22] 0.5498980 0.3694690 0.6364900 0.6306242 0.6126680 0.5760368 0.4579807
#>   [29] 0.4540760 0.5232673 0.5624979 0.6091825 0.4467311 0.4729080 0.6155511
#>   [36] 0.5747637 0.4689331 0.2900026 0.5714818 0.6252113 0.2084648 0.5164600
#>   [43] 0.5895021 0.4781960 0.6219337 0.4698386 0.4870982 0.6131104 0.4987134
#>   [50] 0.5717630 0.6246021 0.6336822 0.3113282 0.6368507 0.5620981 0.5332064
#>   [57] 0.5861844 0.5753732 0.5172882 0.3914588 0.5730957 0.6212242 0.5352548
#>   [64] 0.6166963 0.5826926 0.6194321 0.6346215 0.6360542 0.6337827 0.3677917
#>   [71] 0.6074897 0.4088532 0.6211858 0.3858858 0.6314255 0.6312969 0.5383332
#>   [78] 0.5501453 0.5237341 0.5971225 0.6362928 0.4894178 0.5507644 0.5254038
#>   [85] 0.6256830 0.3799600 0.4169085 0.6101309 0.4104715 0.6279296 0.5463384
#>   [92] 0.5938882 0.6168179 0.5788152 0.6059261 0.6118937 0.5606799 0.3868816
#>   [99] 0.4737257 0.5770828 0.3220908 0.5360758 0.4319531 0.5599972 0.6290184
#>  [106] 0.5668813 0.6205130 0.6293328 0.6272623 0.6122693 0.5587328 0.5049882
#>  [113] 0.3051793 0.6009426 0.5759311 0.5329558 0.4953883 0.4195630 0.3488801
#>  [120] 0.3713825 0.6293937 0.4079712 0.6342412 0.6353557 0.5344720 0.6166060
#>  [127] 0.6145836 0.4787557 0.6159370 0.4435303 0.5680682 0.4134567 0.5604034
#>  [134] 0.6223524 0.6222142 0.4666822 0.4104656 0.6251515 0.5751696 0.3029486
#>  [141] 0.6270005 0.4170419 0.5508251 0.5442412 0.3209490 0.5772917 0.3654247
#>  [148] 0.4395288 0.5472499 0.6357586 0.5314024 0.6346224 0.6219423 0.5127182
#>  [155] 0.4165807 0.5605342 0.6110696 0.5840332 0.6135176 0.2914852 0.5835353
#>  [162] 0.4573072 0.4907642 0.6207798 0.4516017 0.6286756 0.4449222 0.6152135
#>  [169] 0.3954646 0.4889778 0.4554411 0.5491448 0.6173360 0.6249916 0.5291474
#>  [176] 0.4382508 0.5110412 0.5807750 0.3082652 0.5924805 0.6318618 0.5455830
#>  [183] 0.2144941 0.5903940 0.6158348 0.5992869 0.6310108 0.5682647 0.6028657
#>  [190] 0.5014111 0.3321919 0.3237416 0.5741066 0.3406314 0.4787624 0.5116576
#>  [197] 0.6266789 0.5881815 0.6302421 0.4981361 0.5232288 0.6051369 0.6286267
#>  [204] 0.5867073 0.6312247 0.2380819 0.5103474 0.3343464 0.4957874 0.6345263
#>  [211] 0.6073177 0.5310440 0.4946762 0.5352309 0.3151704 0.6353933 0.5233144
#>  [218] 0.6089199 0.6350131 0.4778253 0.3352737 0.3249772 0.5511612 0.2976067
#>  [225] 0.3716696 0.4085868 0.4476639 0.4888044 0.2567917 0.5781288 0.5863188
#>  [232] 0.6333638 0.6349958 0.4497029 0.4184147 0.6126092 0.3763937 0.3140237
#>  [239] 0.5297468 0.4694351 0.6081445 0.5412181 0.6154853 0.6196394 0.6266603
#>  [246] 0.6112887 0.4553350 0.5171945 0.5201911 0.5258431 0.4655982 0.2819886
#>  [253] 0.6282448 0.4215686 0.5548051 0.6110209 0.2804562 0.6155574 0.6321677
#>  [260] 0.5912452 0.6334682 0.5731057 0.4794716 0.5073525 0.3337234 0.4608597
#>  [267] 0.5734342 0.5899181 0.6132093 0.5943552 0.5992753 0.5602455 0.6331631
#>  [274] 0.4283728 0.3656979 0.4413582 0.4907641 0.3340617 0.6220349 0.3498176
#>  [281] 0.5828174 0.3540713 0.5617176 0.6171672 0.5951299 0.4344917 0.6234892
#>  [288] 0.2355929 0.2820776 0.4109798 0.5669006 0.4878998 0.5650190 0.3937041
#>  [295] 0.4946751 0.6162414 0.5884730 0.6117824 0.3905867 0.5665236 0.5751594
#>  [302] 0.5783626 0.5102864 0.5063288 0.6004536 0.4558807 0.5164326 0.5681200
#>  [309] 0.4237661 0.6190042 0.4553182 0.5952796 0.6265730 0.5135872 0.6154363
#>  [316] 0.4490011 0.5627850 0.4934252 0.6073085 0.5964845 0.6264703 0.4959896
#>  [323] 0.6108062 0.2445629 0.5032880 0.4988642 0.5054230 0.5236550 0.5567678
#>  [330] 0.4868169 0.6032521 0.5343491 0.4993182 0.6315328 0.4439322 0.2899944
#>  [337] 0.5459245 0.5503120 0.6328837 0.5254628 0.4907643 0.4871387 0.5274039
#>  [344] 0.5160851 0.4560086 0.5511349 0.6254122 0.4514210 0.4867920 0.5775545
#>  [351] 0.5380659 0.4420309 0.5426239 0.5590395 0.3600087 0.5661078 0.4993523
#>  [358] 0.6310833 0.6109735 0.5393428 0.6211642 0.5030190 0.5828338 0.4349430
#>  [365] 0.5177292 0.4375615 0.3364812 0.5769620 0.6242330 0.5030205 0.2939523
#>  [372] 0.2919352 0.5378008 0.3232712 0.6126469 0.6342668 0.4371585 0.4857195
#>  [379] 0.3250053 0.3841079 0.5657421 0.3640387 0.5130934 0.4707291 0.6341429
#>  [386] 0.5789662 0.6197214 0.6149280 0.4034005 0.6259740 0.6294173 0.4701926
#>  [393] 0.6332121 0.6280026 0.5443160 0.4723348 0.5167603 0.5643454 0.2218265
#>  [400] 0.5186333 0.3003906 0.3438613 0.3851202 0.4712817 0.4408043 0.5192706
#>  [407] 0.6129728 0.4977088 0.5078556 0.4621832 0.6264236 0.6218553 0.6364886
#>  [414] 0.5708970 0.4199439 0.5343389 0.5754435 0.6286242 0.5071244 0.2731176
#>  [421] 0.3588342 0.3524822 0.5265317 0.4690751 0.3075076 0.2981464 0.6040818
#>  [428] 0.5376357 0.6166605 0.6089670 0.3953498 0.4411253 0.3779294 0.2095702
#>  [435] 0.5950563 0.6325626 0.4739803 0.4341733 0.4031456 0.6271136 0.4640386
#>  [442] 0.4204684 0.5244978 0.5896838 0.6341158 0.5777366 0.5810100 0.6310730
#>  [449] 0.5443592 0.4898006 0.3487655 0.3660179 0.3422737 0.4355352 0.6234069
#>  [456] 0.5988559 0.6054003 0.6272093 0.6317477 0.4240549 0.4319856 0.5344767
#>  [463] 0.4847702 0.3086354 0.4863598 0.4843595 0.5501974 0.5829452 0.3565756
#>  [470] 0.6207804 0.4821480 0.5681096 0.6280397 0.2481636 0.4765679 0.6010248
#>  [477] 0.5234882 0.5851540 0.6194848 0.5694812 0.3602561 0.4437520 0.4325012
#>  [484] 0.3363195 0.4177447 0.6306657 0.5204181 0.6263997 0.3763353 0.6310642
#>  [491] 0.2116967 0.4607549 0.6030418 0.6093217 0.6217588 0.5073897 0.4223739
#>  [498] 0.6080279 0.5180106 0.4505512 0.6281868 0.4015603 0.3741296 0.5226196
#>  [505] 0.3727357 0.6316951 0.6341458 0.3506984 0.5189126 0.6182345 0.4860993
#>  [512] 0.5230491 0.5597473 0.3518304 0.5665207 0.6119533 0.6236514 0.3865508
#>  [519] 0.6312434 0.6001366 0.5095290 0.5662660 0.4521175 0.6009226 0.5902575
#>  [526] 0.6366135 0.5133396 0.6335521 0.4319895 0.5544088 0.5504812 0.2173761
#>  [533] 0.4464663 0.6174966 0.6184821 0.5832016 0.6044964 0.6021954 0.2504613
#>  [540] 0.5945694 0.4288782 0.3144497 0.5785801 0.5621867 0.6273348 0.6226980
#>  [547] 0.6332093 0.3712081 0.5006005 0.5906673 0.6278862 0.3681063 0.5976102
#>  [554] 0.5413126 0.3895227 0.5980455 0.6107422 0.6299805 0.4624703 0.5142207
#>  [561] 0.4598335 0.5736639 0.5492279 0.6167352 0.6161164 0.4927961 0.6282874
#>  [568] 0.5953835 0.5759180 0.2594582 0.2372441 0.2862208 0.6273307 0.3855589
#>  [575] 0.6058378 0.6031919 0.3428940 0.6252074 0.5812902 0.6342821 0.4705476
#>  [582] 0.3370678 0.4906443 0.5568022 0.6192486 0.5896636 0.5642246 0.5369756
#>  [589] 0.5756475 0.3033686 0.4692258 0.4796803 0.6198067 0.5964515 0.5795839
#>  [596] 0.6302058 0.6226900 0.5470343 0.4153690 0.5168632 0.4447551 0.4233263
#>  [603] 0.4452254 0.5692898 0.6335120 0.6301772 0.4647665 0.3793293 0.4061856
#>  [610] 0.5289068 0.6080370 0.6303457 0.3116544 0.5997258 0.6162540 0.6068475
#>  [617] 0.6364216 0.6196737 0.3534938 0.2156252 0.5039802 0.4563233 0.3045341
#>  [624] 0.2421332 0.5513591 0.4534947 0.4758278 0.4551949 0.4427932 0.5528196
#>  [631] 0.5917759 0.6271123 0.5893848 0.5003454 0.4305480 0.4985175 0.4656115
#>  [638] 0.6063196 0.3658696 0.4620063 0.5337524 0.4537626 0.6206484 0.4878103
#>  [645] 0.3942528 0.6226156 0.5676665 0.6344001 0.6191047 0.4423194 0.4037743
#>  [652] 0.5591198 0.5238699 0.3113517 0.3943048 0.6310876 0.1601515 0.4526350
#>  [659] 0.2352558 0.4025776 0.4744588 0.3974717 0.6288158 0.4759927 0.6311707
#>  [666] 0.5691453 0.6137133 0.4463501 0.4937083 0.4166530 0.6013743 0.4656406
#>  [673] 0.4196051 0.5041547 0.3977263 0.2927175 0.5139103 0.6336539 0.6339708
#>  [680] 0.3920872 0.6004322 0.5188119 0.6172445 0.6158114 0.4599246 0.5743461
#>  [687] 0.4396942 0.3240908 0.5201655 0.6255877 0.3728242 0.3581155 0.6348057
#>  [694] 0.5633942 0.4245638 0.5810356 0.5603739 0.6253223 0.6077592 0.5941165
#>  [701] 0.6191409 0.5396764 0.4739475 0.4038053 0.5906011 0.5502820 0.5704220
#>  [708] 0.2693134 0.4381253 0.6244780 0.5972424 0.6360789 0.6184275 0.6301819
#>  [715] 0.3368946 0.4461897 0.6021506 0.5420965 0.5782848 0.6212046 0.3650181
#>  [722] 0.6183571 0.3572183 0.5348795 0.4438407 0.3506321 0.4645435 0.6225970
#>  [729] 0.3159927 0.6141073 0.4974583 0.3637649 0.5259179 0.5316568 0.2017726
#>  [736] 0.6277897 0.5289219 0.4836637 0.5295846 0.6240678 0.4903333 0.5279136
#>  [743] 0.5171392 0.3514787 0.5469757 0.6289613 0.6300891 0.5478358 0.6365040
#>  [750] 0.4316076 0.4477987 0.3861417 0.6235274 0.5244979 0.3555305 0.5199921
#>  [757] 0.5687701 0.4349287 0.4375990 0.6320334 0.5795257 0.4391309 0.6207192
#>  [764] 0.5268319 0.5689055 0.3814241 0.5961305 0.5864450 0.5589907 0.3785462
#>  [771] 0.3316558 0.5324322 0.5791553 0.4950030 0.6310966 0.3960935 0.5487231
#>  [778] 0.6162800 0.5468457 0.6301377 0.5995084 0.6228970 0.5401552 0.5631466
#>  [785] 0.6127445 0.5816185 0.4899236 0.2738833 0.5013999 0.6284455 0.6337520
#>  [792] 0.6048169 0.5918516 0.6292261 0.4592504 0.5718666 0.5635348 0.5854236
#>  [799] 0.6221509 0.5449255 0.5285729 0.5627531 0.5806874 0.3841315 0.6361135
#>  [806] 0.4953878 0.2667858 0.3993132 0.2882159 0.4968648 0.6374088 0.5240189
#>  [813] 0.5362542 0.4707712 0.6339191 0.6041877 0.6267897 0.4756080 0.5976727
#>  [820] 0.6260212 0.4194103 0.6284446 0.6350910 0.6087920 0.6356300 0.5958984
#>  [827] 0.6310163 0.3259878 0.5644550 0.4025124 0.4647859 0.3350220 0.4787996
#>  [834] 0.5147372 0.4605675 0.4060274 0.3340625 0.6292617 0.6265776 0.3932857
#>  [841] 0.3936139 0.6181260 0.2751822 0.5248792 0.6232512 0.5162159 0.5780396
#>  [848] 0.4034054 0.5181452 0.4807423 0.6316876 0.6334858 0.6310610 0.4839599
#>  [855] 0.5750939 0.5370777 0.5093767 0.5120306 0.5233530 0.6354201 0.2962075
#>  [862] 0.5782383 0.5683217 0.4611922 0.5993442 0.6299565 0.3967887 0.5636015
#>  [869] 0.3589966 0.6286638 0.3697711 0.3839319 0.5386242 0.4775034 0.3151843
#>  [876] 0.3848272 0.5199583 0.3576236 0.3771404 0.5188192 0.6332005 0.4791474
#>  [883] 0.6147499 0.5578307 0.6313850 0.6328016 0.6224536 0.5193310 0.4795770
#>  [890] 0.6086293 0.5900342 0.4970858 0.4662966 0.3162165 0.6176077 0.5701661
#>  [897] 0.4242119 0.6240008 0.6042503 0.6096156 0.5053933 0.5363969 0.5835724
#>  [904] 0.3125335 0.5044611 0.5237147 0.4798265 0.5513145 0.6316295 0.3722557
#>  [911] 0.3716820 0.6111699 0.3147572 0.1665517 0.6008452 0.3884744 0.2472651
#>  [918] 0.6297467 0.5196551 0.6048465 0.5976123 0.4102202 0.5669134 0.3150132
#>  [925] 0.4263606 0.5836956 0.4205462 0.4141819 0.6175562 0.6280492 0.5094887
#>  [932] 0.3006627 0.2339602 0.2403166 0.4235083 0.5410465 0.3295207 0.6274547
#>  [939] 0.5640970 0.4055716 0.6021384 0.2571944 0.5019379 0.6317557 0.4388105
#>  [946] 0.6352481 0.6037116 0.4928955 0.4744770 0.4112464 0.6221252 0.5744561
#>  [953] 0.5196747 0.5177324 0.4725303 0.5587121 0.5585537 0.1482683 0.5471279
#>  [960] 0.5538261 0.6132567 0.5714377 0.4901121 0.4060362 0.5639523 0.6348961
#>  [967] 0.4908297 0.6177689 0.3039919 0.5438748 0.4519374 0.6351808 0.6227518
#>  [974] 0.5211898 0.3077041 0.5175888 0.5463321 0.3954046 0.3008286 0.6282459
#>  [981] 0.3172420 0.5813043 0.4692551 0.5616514 0.4113244 0.6319068 0.4315120
#>  [988] 0.6232941 0.3274399 0.4196015 0.5112502 0.5636144 0.5357319 0.5165005
#>  [995] 0.4347852 0.5814915 0.5295942 0.6302892 0.5095274 0.4470930 0.6220685
#> [1002] 0.6031638 0.5949167 0.6332395 0.5989503 0.4453559 0.6328399 0.5651982
#> [1009] 0.6081813 0.5669768 0.5472535 0.5261336 0.5440786 0.6110058 0.5082158
#> [1016] 0.4718325 0.5522990 0.5560969 0.4491829 0.5057780 0.6105554 0.4780434
#> [1023] 0.6116319 0.6038185 0.3424305 0.3603820 0.2130853 0.3988120 0.2610077
#> [1030] 0.3814498 0.3731425 0.6180976 0.3364811 0.6106703 0.5128430 0.6151940
#> [1037] 0.5981454 0.5689511 0.6093419 0.3599769 0.5463804 0.6335989 0.4059631
#> [1044] 0.5331885 0.4484582 0.5590002 0.6327802 0.5436614 0.5495525 0.5807634
#> [1051] 0.6128090 0.2821544 0.5573909 0.4761125 0.5243067 0.4863942 0.6291943
#> [1058] 0.5738120 0.5922975 0.4232084 0.6236662 0.4973884 0.4038657 0.5869718
#> [1065] 0.6191795 0.6176250 0.4489893 0.5195959 0.5209997 0.5380163 0.6338627
#> [1072] 0.6290501 0.4816923 0.4699936 0.3456574 0.4313712 0.5583240 0.3774122
#> [1079] 0.5187069 0.3385637 0.5766082 0.4384961 0.4563241 0.6331067 0.3736198
#> [1086] 0.5947670 0.4947165 0.5546529 0.4455737 0.5281409 0.6302632 0.4240166
#> [1093] 0.6301046 0.4300323 0.6331033 0.5411014 0.4835477 0.4245399 0.5643302
#> [1100] 0.4911515 0.3679446 0.6028067 0.5493206 0.5818494 0.4996365 0.5383897
#> [1107] 0.6296564 0.5610426 0.4290313 0.5697876 0.4876329 0.5334303 0.6214045
#> [1114] 0.6242321 0.5279250 0.3855861 0.6086961 0.6281856 0.3786901 0.6059679
#> [1121] 0.2535309 0.6340306 0.5594654 0.6125931 0.4160740 0.5635050 0.4950759
#> [1128] 0.4331312 0.4884424 0.5715263 0.3458089 0.3655642 0.5036871 0.6232775
#> [1135] 0.2499420 0.5653807 0.6155403 0.5865035 0.2647364 0.1980682 0.5651188
#> [1142] 0.6263170 0.5027793 0.4218357 0.3580470 0.6171854 0.5077206 0.5458805
#> [1149] 0.6259842 0.6297200 0.4418360 0.6227358 0.5706768 0.6196949 0.6354399
#> [1156] 0.4550147 0.5848764 0.4938661 0.5647317 0.4791298 0.4863487 0.6065359
#> [1163] 0.6339042 0.6291448 0.6331573 0.2358555 0.5680450 0.4994313 0.3529505
#> [1170] 0.6146770 0.6317057 0.3084176 0.5801749 0.3492815 0.5661407 0.5810523
#> [1177] 0.4668322 0.5739634 0.4232758 0.5200966 0.6164735 0.6312811 0.5540664
#> [1184] 0.4969293 0.4102588 0.4164426 0.5657184 0.5887773 0.4635443 0.2040928
#> [1191] 0.5754663 0.3152714 0.5703626 0.4391808 0.6312118 0.4716953 0.5158079
#> [1198] 0.6142175 0.4995656 0.6353242 0.6167455 0.5114181 0.4269275 0.6241194
#> [1205] 0.5978878 0.5409567 0.3522448 0.6350448 0.5440234 0.3962578 0.5740438
#> [1212] 0.2733090 0.6221000 0.5441683 0.5600057 0.3422381 0.3836824 0.6247980
#> [1219] 0.5252466 0.4477855 0.4956362 0.6065348 0.3377301 0.5913681 0.5682318
#> [1226] 0.3786343 0.3460017 0.2519237 0.6042224 0.5471976 0.3315589 0.4970374
#> [1233] 0.5816660 0.4974054 0.4294619 0.5675367 0.6202194 0.6343995 0.5662340
#> [1240] 0.3066187 0.5884776 0.5597429 0.4816609 0.5456299 0.5817454 0.2912314
#> [1247] 0.4911649 0.6104203 0.6147436 0.6368699 0.3366291 0.5475173 0.4425569
#> [1254] 0.4999441 0.5678319 0.4872551 0.5613642 0.4713946 0.5598935 0.6357296
#> [1261] 0.4043941 0.6261498 0.4924055 0.6160237 0.4680970 0.6367289 0.5177453
#> [1268] 0.6283802 0.6282151 0.2980256 0.5299257 0.4584694 0.2777927 0.6146265
#> [1275] 0.6178620 0.5350980 0.6334681 0.6077421 0.3503879 0.5926009 0.5645429
#> [1282] 0.4508415 0.6240619 0.5530775 0.5230861 0.6184134 0.4504311 0.5232912
#> [1289] 0.1651927 0.4719406 0.4932092 0.5911109 0.5244351 0.5634940 0.5780053
#> [1296] 0.5169906 0.3957591 0.4517691 0.5402785 0.5437097 0.5212200 0.5487128
#> [1303] 0.4910608 0.5330297 0.4387320 0.4304743 0.3845018 0.4730081 0.5163181
#> [1310] 0.5078072 0.6315831 0.5080045 0.5173042 0.3910029 0.6239505 0.5397493
#> [1317] 0.5832833 0.5870071 0.4363159 0.3533933 0.5290850 0.6120671 0.6217367
#> [1324] 0.5957055 0.6175114 0.6268639 0.3341747 0.3015113 0.6184139 0.5833168
#> [1331] 0.4767627 0.4209834 0.4351852 0.4955233 0.3594634 0.5898042 0.6045349
#> [1338] 0.5635497 0.4628321 0.4604791 0.6024431 0.2893890 0.5749768 0.4549575
#> [1345] 0.6313493 0.3366260 0.5007445 0.5854605 0.6124005 0.5670756 0.6338317
#> [1352] 0.6101628 0.6047587 0.6041409 0.3889395 0.6102488 0.5077943 0.5087929
#> [1359] 0.4704248 0.6181156 0.6110793 0.5851147 0.4507213 0.2697361 0.6137332
#> [1366] 0.6202260 0.4233934 0.6143848 0.5960379 0.5857240 0.4136331 0.6291838
#> [1373] 0.5467662 0.6348779 0.3898125 0.4807697 0.4792792 0.3733325 0.3052003
#> [1380] 0.6277508 0.3236390 0.4026692 0.6101662 0.4783569 0.6293727 0.5298125
#> [1387] 0.4163073 0.4357026 0.4492658 0.6083922 0.4626191 0.6062209 0.4116020
#> [1394] 0.5794290 0.4331861 0.5376289 0.6303999 0.4395190 0.5786034 0.5793440
#> [1401] 0.6304945 0.6315934 0.6288932 0.5577008 0.6259892 0.6159478 0.4838799
#> [1408] 0.5311253 0.5547327 0.5473493 0.4343987 0.6369104 0.5792322 0.6081168
#> [1415] 0.5559827 0.6127285 0.6329238 0.4449173 0.4246893 0.4837837 0.6306067
#> [1422] 0.5853764 0.6231562 0.4503533 0.4414939 0.5782582 0.4457541 0.6002144
#> [1429] 0.3549352 0.4954533 0.4846741 0.4300163 0.6302884 0.6065819 0.4723503
#> [1436] 0.6302894 0.2460225 0.6239235 0.3783983 0.6127092 0.6034642 0.3289884
#> [1443] 0.2807099 0.4588279 0.6363084 0.5725049 0.2593765 0.5299879 0.6221083
#> [1450] 0.5913924 0.6274378 0.3755199 0.5202661 0.5445315 0.3752770 0.5868225
#> [1457] 0.5711651 0.5172068 0.5071115 0.4207703 0.5634899 0.6130075 0.5765701
#> [1464] 0.6151693 0.5095900 0.6139315 0.6308729 0.4339944 0.4129853 0.5888326
#> [1471] 0.4726393 0.5556919 0.3708281 0.2586952 0.3328556 0.2889363 0.5516927
#> [1478] 0.4918340 0.6324149 0.5737921 0.4425483 0.4149613 0.4739838 0.3870572
#> [1485] 0.5987156 0.5463849 0.5333943 0.6172436 0.4433606 0.5063457 0.4399224
#> [1492] 0.5404101 0.5915000 0.6156019 0.5626383 0.5053854 0.6209221 0.4434433
#> [1499] 0.6100181 0.6086059 0.4983855 0.6231707 0.4794264 0.4589483 0.5932648
#> [1506] 0.4110069 0.4176181 0.5350045 0.4465064 0.5387907 0.5327294 0.4038936
#> [1513] 0.5854686 0.5314150 0.3039855 0.3259169 0.6062543 0.6329590 0.6022887
#> [1520] 0.6314046 0.6075567 0.6164928 0.2751976 0.6300722 0.4074884 0.6244816
#> [1527] 0.6278040 0.4949864 0.6320611 0.4979222 0.5801825 0.2875041 0.2758681
#> [1534] 0.6357923 0.6334576 0.5479338 0.6366539 0.3911373 0.4749977 0.5985033
#> [1541] 0.4022466 0.5428029 0.4865894 0.6332584 0.4959775 0.6076415 0.6067614
#> [1548] 0.6024831 0.5056661 0.5700491 0.5236822 0.5953677 0.5667715 0.6158365
#> [1555] 0.5728074 0.5469095 0.4852725 0.4802223 0.5956806 0.5391064 0.6218521
#> [1562] 0.4098597 0.6034631 0.3450600 0.4125250 0.4606107 0.6264731 0.5123338
#> [1569] 0.5218297 0.6136100 0.3912692 0.6279591 0.3259670 0.6117305 0.6012035
#> [1576] 0.6215874 0.3402441 0.5547300 0.5453705 0.3925839 0.1948972 0.4552859
#> [1583] 0.5687396 0.3960977 0.4028870 0.2387630 0.5719240 0.3310889 0.3454836
#> [1590] 0.4356716 0.5909345 0.2985915 0.3560170 0.6339002 0.6282742 0.3376678
#> [1597] 0.5295566 0.6365861 0.4750434 0.5289899 0.5775289 0.3784653 0.4701706
#> [1604] 0.2963073 0.4474272 0.5341741 0.6046495 0.4778614 0.5353823 0.4741635
#> [1611] 0.5044810 0.5436225 0.6362432 0.4612441 0.5707948 0.2933015 0.3550865
#> [1618] 0.5506143 0.3620881 0.4737120 0.5117952 0.6088439 0.4941354 0.6274679
#> [1625] 0.6364444 0.5827827 0.4556713 0.5745136 0.6227194 0.5102415 0.6150061
#> [1632] 0.6054406 0.4983483 0.3636911 0.6101884 0.5860757 0.5232158 0.5146077
#> [1639] 0.4046631 0.4465804 0.6106466 0.2373987 0.6231667 0.4778329 0.3437171
#> [1646] 0.2980784 0.2858226 0.6329139 0.6258149 0.4482006 0.4881872 0.2375925
#> [1653] 0.5862133 0.4355302 0.6359366 0.5663658 0.4073282 0.4629336 0.6251575
#> [1660] 0.6113839 0.5097586 0.5829083 0.5986968 0.3414249 0.5994151 0.6054066
#> [1667] 0.4485668 0.5769245 0.5496496 0.5396149 0.6369101 0.6091950 0.4772047
#> [1674] 0.5656796 0.5475357 0.3847368 0.5097539 0.5805283 0.5000006 0.6151204
#> [1681] 0.6051149 0.3627772 0.5923451 0.5458814 0.2755551 0.6338619 0.5769662
#> [1688] 0.5650207 0.5732220 0.5559323 0.5904275 0.4505184 0.2769967 0.5877175
#> [1695] 0.4490051 0.2869037 0.5444259 0.5830838 0.6336507 0.5229710 0.4628339
#> [1702] 0.4621400 0.5179274 0.5328115 0.5736484 0.3765249 0.3592438 0.6097592
#> [1709] 0.6124922 0.6368414 0.6343071 0.3991913 0.5717661 0.4978004 0.5201082
#> [1716] 0.5749761 0.6348722 0.5738563 0.6106312 0.5943223 0.5071991 0.5888312
#> [1723] 0.5725018 0.6116546 0.6114338 0.4143200 0.5095289 0.6337253 0.5617567
#> [1730] 0.5486894 0.3703925 0.3851718 0.5442044 0.4475834 0.6246299 0.5791722
#> [1737] 0.5500823 0.5284372 0.6260059 0.6004079 0.3268480 0.6314821 0.5902585
#> [1744] 0.5461369 0.6324324 0.3431131 0.5046840 0.6271111 0.5429758 0.4375269
#> [1751] 0.6287372 0.4521991 0.6227449 0.4833830 0.5576862 0.4618899 0.4730532
#> [1758] 0.4448125 0.6328549 0.4489654 0.5182546 0.4367473 0.5303005 0.5706547
#> [1765] 0.5913444 0.4733216 0.4300171 0.6115821 0.4850984 0.4484754 0.6166955
#> [1772] 0.3346978 0.6178947 0.5304771 0.4277000 0.5507006 0.6120059 0.5093960
#> [1779] 0.4163354 0.4488826 0.3450031 0.5383146 0.4497813 0.3932125 0.4223870
#> [1786] 0.4894414 0.6098642 0.5662779 0.5758614 0.2386097 0.4074268 0.5893482
#> [1793] 0.4135044 0.2729651 0.5104412 0.5082557 0.6311420 0.6366434 0.6273372
#> [1800] 0.4180961 0.5421409 0.5041993 0.4513640 0.4824784 0.6299934 0.6244223
#> [1807] 0.4434380 0.3202608 0.6209802 0.4923363 0.6298663 0.2946075 0.3467536
#> [1814] 0.5363158 0.6230108 0.3341932 0.5193219 0.6306042 0.5677416 0.3684988
#> [1821] 0.4261291 0.4772389 0.5520407 0.5648834 0.5018817 0.5434187 0.5048362
#> [1828] 0.5912415 0.2467870 0.6202248 0.3788475 0.6122489 0.3876368 0.6247276
#> [1835] 0.4213392 0.5593816 0.4254220 0.6330394 0.3918311 0.2481672 0.6256293
#> [1842] 0.5159081 0.3867396 0.3804653 0.4711431 0.5737492 0.4909460 0.4964397
#> [1849] 0.5671162 0.5293039 0.5494225 0.6362760 0.6281094 0.6360754 0.3914221
#> [1856] 0.2947310 0.5439916 0.5208415 0.5548876 0.4545840 0.5742372 0.5721513
#> [1863] 0.6263338 0.3505576 0.3236886 0.4613931 0.4331720 0.5715350 0.5788099
#> [1870] 0.5361155 0.3250927 0.3946843 0.2321738 0.6192952 0.4727593 0.5136741
#> [1877] 0.3684589 0.6020784 0.5662950 0.2313513 0.5735674 0.4815609 0.4555896
#> [1884] 0.6316465 0.5962367 0.4251114 0.5779292 0.3568041 0.6308946 0.3727338
#> [1891] 0.5881035 0.5079567 0.3856456 0.2486141 0.5007128 0.5957284 0.6145475
#> [1898] 0.5634596 0.4325275 0.5210302 0.5228863 0.5949880 0.3689048 0.4091105
#> [1905] 0.6204346 0.4848066 0.3127370 0.4355447 0.5146242 0.6084388 0.6049890
#> [1912] 0.1723415 0.6273386 0.4811191 0.4521040 0.6051712 0.5387527 0.6233448
#> [1919] 0.3821215 0.2916643 0.5595185 0.5467285 0.5658900 0.3263173 0.4134406
#> [1926] 0.5523290 0.5619793 0.3995948 0.3634039 0.4799382 0.5501985 0.4901997
#> [1933] 0.6319851 0.4878586 0.5291608 0.2302776 0.2082573 0.2585341 0.4439692
#> [1940] 0.5880746 0.6227444 0.3952460 0.6288231 0.6305001 0.4640879 0.5451295
#> [1947] 0.6354243 0.6339354 0.6341320 0.5361745 0.4993715 0.5440455 0.5289243
#> [1954] 0.4013168 0.5634083 0.3410615 0.6362714 0.5831030 0.4705204 0.4324600
#> [1961] 0.3646531 0.5511288 0.5281042 0.3706768 0.4644783 0.4738044 0.5187857
#> [1968] 0.3116701 0.4117601 0.6251857 0.5095295 0.4580082 0.5848267 0.6061144
#> [1975] 0.5911666 0.6369118 0.6319910 0.4011942 0.5563856 0.3983621 0.3940644
#> [1982] 0.6369117 0.4995204 0.5696294 0.4741151 0.6339787 0.5302811 0.5375587
#> [1989] 0.5505962 0.4429196 0.2524450 0.2321666 0.5868770 0.3314997 0.3145970
#> [1996] 0.4007894 0.5652154 0.4837732 0.5859580 0.4361609 0.5064799 0.6347217
#> [2003] 0.6227972 0.3012076 0.6036934 0.6117177 0.2851746 0.4228363 0.3755303
#> [2010] 0.4732107 0.4589819 0.3569898 0.4433948 0.6322800 0.5692034 0.5738599
#> [2017] 0.4987185 0.4023199 0.3304653 0.6288918 0.3310831 0.4746252 0.6118700
#> [2024] 0.6144053 0.6287483 0.4237662 0.5543769 0.2806594 0.5038787 0.6195126
#> [2031] 0.5377773 0.5928491 0.2932912 0.4103788 0.3109684 0.3133274 0.4967437
#> [2038] 0.4487670 0.6289806 0.5902683 0.5427936 0.4742070 0.5678579 0.6206714
#> [2045] 0.4539283 0.4681978 0.4961873 0.5775472 0.6360627 0.4065393 0.5519910
#> [2052] 0.5570618 0.4942112 0.4860479 0.4822942 0.6126897 0.4541292 0.4225535
#> [2059] 0.5250258 0.6292396 0.6228170 0.4852535 0.6345251 0.5404944 0.3822136
#> [2066] 0.4410310 0.6170755 0.5832175 0.5191228 0.5314273 0.5868544 0.4887063
#> [2073] 0.3515874 0.4828378 0.4778515 0.6319832 0.3463711 0.5272182 0.6314387
#> [2080] 0.3012434 0.3618968 0.5942230 0.4874214 0.6063214 0.6337434 0.6269144
#> [2087] 0.5840594 0.3784778 0.2469337 0.5194991 0.5351787 0.6287519 0.5037145
#> [2094] 0.4016883 0.6189749 0.4021768 0.5624962 0.5776763 0.5191390 0.6130047
#> [2101] 0.5020763 0.5209178 0.5879167 0.3925898 0.6299180 0.6214900 0.6317462
#> [2108] 0.4311937 0.6097725 0.5217210 0.5768689 0.6247752 0.6049423 0.6369104
#> [2115] 0.4807412 0.6301290 0.4628415 0.6308341 0.2622966 0.3483568 0.5818100
#> [2122] 0.4559683 0.6327942 0.5879156 0.3396496 0.6283957 0.3182979 0.5453091
#> [2129] 0.6300495 0.3594730 0.6100903 0.4159903 0.5385420 0.3924541 0.4183151
#> [2136] 0.2893739 0.5954218 0.5115528 0.3707398 0.3571383 0.3661262 0.5935252
#> [2143] 0.4341701 0.5529606 0.5959270 0.5472066 0.5970878 0.4557619 0.1923282
#> [2150] 0.6007146 0.5809356 0.6090224 0.5198454 0.4780343 0.4391805 0.2254765
#> [2157] 0.3841535 0.6233053 0.6339893 0.5696094 0.6010553 0.6282085 0.6079343
#> [2164] 0.6297937 0.5976361 0.3830892 0.4269227 0.6333691 0.5846759 0.5848890
#> [2171] 0.6166328 0.2837895 0.6334741 0.6360419 0.5108507 0.2738298 0.5216355
#> [2178] 0.5288119 0.6083192 0.6213782 0.3753430 0.5932443 0.5392404 0.4066513
#> [2185] 0.6191503 0.6083239 0.5243459 0.5468611 0.5401287 0.5696813 0.4931612
#> [2192] 0.4775665 0.5727676 0.4673134 0.3555740 0.4436337 0.5139743 0.6352612
#> [2199] 0.2805802 0.6049996 0.5547151 0.5263304 0.6065139 0.4961179 0.5934144
#> [2206] 0.5546773 0.6279399 0.3231651 0.3890734 0.3387102 0.6252698 0.5242038
#> [2213] 0.5672344 0.3458794 0.3828955 0.5929022 0.6326202 0.5605498 0.4327073
#> [2220] 0.4457796 0.4326442 0.5262834 0.2624742 0.4461961 0.3638217 0.5951550
#> [2227] 0.4970484 0.4807836 0.4638635 0.5996850 0.5210776 0.5681612 0.3698503
#> [2234] 0.3184645 0.5282766 0.4324582 0.5292722 0.5409320 0.5344604 0.5532366
#> [2241] 0.5310692 0.6239652 0.4450339 0.6177145 0.5333713 0.4597942 0.5821255
#> [2248] 0.6337207 0.5909418 0.4663344 0.5934816 0.5907450 0.4595873 0.5436927
#> [2255] 0.5590376 0.4160703 0.5414792 0.4165471 0.5746663 0.5360711 0.5391391
#> [2262] 0.5743376 0.3439386 0.6123682 0.4694450 0.3076214 0.5676408 0.5869735
#> [2269] 0.4927377 0.3013651 0.4034731 0.3359507 0.2380541 0.3329657 0.5333436
#> [2276] 0.5803856 0.6328274 0.6106680 0.6275885 0.5789455 0.6103579 0.6073277
#> [2283] 0.4507212 0.5681878 0.5194633 0.3372992 0.4741272 0.6156013 0.4906184
#> [2290] 0.5996650 0.4674452 0.3412221 0.5326153 0.3994207 0.5959641 0.5958606
#> [2297] 0.5750083 0.5756826 0.4507846 0.4528732 0.6184321 0.4549856 0.3732254
#> [2304] 0.6189336 0.6142770 0.5932899 0.3837531 0.3790425 0.4530990 0.5761721
#> [2311] 0.3818319 0.3978668 0.4276280 0.5538897 0.5691858 0.5106542 0.4167063
#> [2318] 0.5960120 0.5983022 0.3263296 0.4477555 0.6298218 0.6360168 0.5070046
#> [2325] 0.5634675 0.4686200 0.4893755 0.6312498 0.6197505 0.6116478 0.6030761
#> [2332] 0.5739665 0.5168780 0.6003201 0.3844856 0.3014954 0.6079419 0.4479261
#> [2339] 0.1752431 0.4173688 0.4793526 0.4975026 0.6294932 0.4725288 0.6334579
#> [2346] 0.4411683 0.5636036 0.4505879 0.6369121 0.5455522 0.3703501 0.4668209
#> [2353] 0.5235060 0.4246854 0.3670220 0.4271060 0.3247419 0.6260862 0.3135832
#> [2360] 0.3697522 0.5842192 0.5733173 0.6052641 0.5034475 0.5704042 0.2531564
#> [2367] 0.4801187 0.5992792 0.6280079 0.6300788 0.6178498 0.6004557 0.5301442
#> [2374] 0.5800974 0.4047388 0.5636572 0.5764446 0.6335244 0.6175934 0.6129214
#> [2381] 0.5346288 0.5329888 0.4111991 0.6079685 0.4900683 0.2430982 0.6151539
#> [2388] 0.6267827 0.6329516 0.6345881 0.4023209 0.6326061 0.3661858 0.2973182
#> [2395] 0.4091828 0.5485310 0.5714551 0.5112102 0.4918853 0.2469303 0.5451783
#> [2402] 0.6097631 0.5773849 0.6262696 0.4940098 0.5959123 0.6369114 0.6303773
#> [2409] 0.5357349 0.6146599 0.6283412 0.3402034 0.5277221 0.6186530 0.5520492
#> [2416] 0.5083863 0.5410161 0.4800619 0.3862877 0.6057120 0.6072873 0.5282701
#> [2423] 0.6345839 0.6049354 0.5851882 0.5620542 0.5306932 0.5536279 0.6286045
#> [2430] 0.5013074 0.4954693 0.1296743 0.3568171 0.6274646 0.4435893 0.6186819
#> [2437] 0.5970245 0.4405927 0.4326852 0.6367698 0.3906594 0.4666474 0.5930445
#> [2444] 0.6127242 0.5058575 0.6108014 0.4421002 0.4050400 0.2318119 0.2529767
#> [2451] 0.5041189 0.4249185 0.5908797 0.6234235 0.6205631 0.4525042 0.4885252
#> [2458] 0.4397303 0.5609053 0.5694694 0.6200868 0.5440852 0.5573632 0.4971038
#> [2465] 0.4242272 0.5339350 0.5060000 0.5207749 0.3597484 0.5040106 0.5195806
#> [2472] 0.5211807 0.6230454 0.5631250 0.4887984 0.4557663 0.4425290 0.5136242
#> [2479] 0.4209042 0.5351488 0.5756508 0.4096971 0.5455467 0.6295393 0.5512333
#> [2486] 0.3386097 0.6302039 0.6298676 0.2325041 0.5842174 0.3681130 0.3910299
#> [2493] 0.5885426 0.5748600 0.3956928 0.5371274 0.4206052 0.6290399 0.6287648
#> [2500] 0.4090970 0.6264661 0.4158932 0.4981842 0.5267906 0.3991386 0.4725953
#> [2507] 0.5985990 0.4068447 0.5753119 0.5356435 0.3930476 0.5387261 0.4377587
#> [2514] 0.5488338 0.5280347 0.4194232 0.6311822 0.4517013 0.5640121 0.6307841
#> [2521] 0.5324711 0.6306023 0.6367774 0.4568374 0.5290648 0.4773242 0.4636563
#> [2528] 0.3657952 0.5639571 0.5068316 0.6146256 0.1183158 0.6338579 0.4793052
#> [2535] 0.3070355 0.2972505 0.5502093 0.5312642 0.2292877 0.4320590 0.3185749
#> [2542] 0.3330568 0.3971836 0.5234296 0.4840244 0.3790701 0.6026756 0.4905182
#> [2549] 0.6369104 0.5817933 0.5532795 0.6187556 0.5570200 0.5747724 0.5783315
#> [2556] 0.4106473 0.5848723 0.6254434 0.3648136 0.3237824 0.5828630 0.4025965
#> [2563] 0.3382024 0.5606057 0.3377968 0.5373758 0.4126391 0.2630703 0.6294931
#> [2570] 0.5789702 0.6185747 0.5421189 0.2769453 0.5320772 0.6283783 0.6106699
#> [2577] 0.4269329 0.5047696 0.3053062 0.4446974 0.3992325 0.5508121 0.3360157
#> [2584] 0.5293793 0.2987328 0.5407744 0.3580119 0.6291124 0.5807258 0.3976632
#> [2591] 0.5988855 0.5674082 0.5850040 0.5811616 0.1986274 0.6302644 0.6046180
#> [2598] 0.5525753 0.6292963 0.5854905 0.6129663 0.6322605 0.5878164 0.5438499
#> [2605] 0.5871576 0.6311866 0.5891719 0.4890591 0.3994751 0.6352107 0.6195683
#> [2612] 0.6292244 0.6180750 0.2172640 0.4014851 0.6332954 0.6282122 0.6162243
#> [2619] 0.4412370 0.4544924 0.6211449 0.5650071 0.3293091 0.6272482 0.5144913
#> [2626] 0.6063348 0.6177201 0.4958109 0.4334473 0.3053450 0.2104625 0.5760918
#> [2633] 0.6192844 0.5476324 0.5597003 0.4361380 0.4739761 0.4572406 0.3928825
#> [2640] 0.4897656 0.4025181 0.6078775 0.5260185 0.5705100 0.6256784 0.5993974
#> [2647] 0.6109825 0.1902332 0.6327203 0.5421999 0.5398597 0.5953139 0.6305565
#> [2654] 0.5404123 0.6350244 0.5187810 0.5180567 0.5486229 0.5282529 0.5231510
#> [2661] 0.6213744 0.6296019 0.6255491 0.5824730 0.4207166 0.3023637 0.4744279
#> [2668] 0.5668855 0.5003126 0.4479076 0.6337249 0.4727547 0.6347175 0.5705907
#> [2675] 0.2652082 0.3968999 0.5466300 0.6308326 0.3987328 0.4836801 0.3874354
#> [2682] 0.5694731 0.3154031 0.4461491 0.6213020 0.3144496 0.6064576 0.6107715
#> [2689] 0.5625425 0.6141063 0.6262874 0.6324897 0.6236260 0.6286904 0.5993488
#> [2696] 0.5578118 0.4176593 0.5044868 0.5861091 0.5987339 0.6093295 0.4844020
#> [2703] 0.5835780 0.3852581 0.5930528 0.6342541 0.5132954 0.3569009 0.4635568
#> [2710] 0.4512326 0.5630167 0.6233492 0.3736552 0.6020091 0.3702003 0.5001731
#> [2717] 0.5905284 0.4954797 0.6136347 0.5565605 0.4523359 0.5847440 0.5774688
#> [2724] 0.5659042 0.5879213 0.4153004 0.5798287 0.4868420 0.3739759 0.2596463
#> [2731] 0.4224089 0.5218191 0.5995747 0.6144626 0.6349951 0.3038381 0.5186097
#> [2738] 0.5947718 0.6030764 0.4949484 0.5769756 0.5508160 0.4489568 0.6078776
#> [2745] 0.6002827 0.5006119 0.4936641 0.5586604 0.5901596 0.6304154 0.5298671
#> [2752] 0.4146582 0.5993610 0.6057446 0.5479114 0.2142894 0.5873390 0.3537759
#> [2759] 0.6287856 0.4442948 0.2947190 0.6296066 0.3539226 0.3481909 0.4009069
#> [2766] 0.5813467 0.5504535 0.6133799 0.3725793 0.3056858 0.4681958 0.4556152
#> [2773] 0.6366619 0.3709997 0.2074070 0.2181194 0.4918079 0.4319285 0.4538844
#> [2780] 0.6037141 0.4146886 0.5743026 0.6295366 0.5377702 0.5368800 0.6208881
#> [2787] 0.6179033 0.4486323 0.3559900 0.3242083 0.5824220 0.4633715 0.6044654
#> [2794] 0.6288951 0.4612110 0.6369104 0.6127666 0.5748211 0.4525072 0.6084375
#> [2801] 0.5090347 0.5356710 0.5143657 0.6038177 0.6003125 0.5193526 0.6225299
#> [2808] 0.4741061 0.3163946 0.4017794 0.5699610 0.3027305 0.6232449 0.4736269
#> [2815] 0.6234543 0.4203876 0.4407040 0.6354004 0.4625114 0.4163435 0.6138276
#> [2822] 0.5654726 0.5627066 0.4218920 0.6219763 0.4306435 0.4711452 0.6243598
#> [2829] 0.4397251 0.5071188 0.4242153 0.5399606 0.5569049 0.6369104 0.4907648
#> [2836] 0.4521338 0.5681170 0.5589135 0.4416640 0.6314006 0.6027247 0.6256437
#> [2843] 0.6065510 0.5951922 0.3970645 0.6144318 0.6323080 0.4597710 0.2996694
#> [2850] 0.5923468 0.3823868 0.4746171 0.5587622 0.4883670 0.6239699 0.5055271
#> [2857] 0.3233353 0.3289026 0.6348628 0.5175606 0.3097860 0.5682564 0.5171356
#> [2864] 0.5263602 0.4961334 0.6355274 0.4899627 0.5587663 0.6364835 0.6356724
#> [2871] 0.5896975 0.5628347 0.6115363 0.3166358 0.3925645 0.4944741 0.3410303
#> [2878] 0.3986425 0.5596451 0.6210366 0.4963722 0.5400798 0.5032985 0.4194726
#> [2885] 0.6317329 0.5158437 0.4630108 0.6327369 0.5136516 0.5448533 0.4352158
#> [2892] 0.4496188 0.4631308 0.6045509 0.3904313 0.2384649 0.5567492 0.3982299
#> [2899] 0.6115762 0.4132332 0.1923618 0.3716462 0.4327555 0.5582756 0.4044005
#> [2906] 0.5286098 0.3085107 0.5557085 0.4987523 0.6229178 0.4430531 0.3789448
#> [2913] 0.4079438 0.4071564 0.6342735 0.4998520 0.5685968 0.6255736 0.6202019
#> [2920] 0.6123649 0.6343135 0.2888582 0.6051331 0.5502520 0.5386896 0.3877998
#> [2927] 0.5855333 0.5042747 0.5173238 0.6208300 0.5148231 0.5538765 0.6354517
#> [2934] 0.6342246 0.3621145 0.6302574 0.6147643 0.6269317 0.4850645 0.6165221
#> [2941] 0.5225207 0.5655849 0.5068184 0.4243365 0.6308296 0.5344123 0.5213505
#> [2948] 0.4093292 0.6311830 0.5404244 0.2934971 0.5384163 0.4321566 0.1719303
#> [2955] 0.3438458 0.5928890 0.2673648 0.4275929 0.5135556 0.5962281 0.6096471
#> [2962] 0.6088719 0.3137894 0.6248802 0.5748367 0.6264698 0.5111425 0.4717990
#> [2969] 0.5564476 0.5070722 0.6103059 0.6241498 0.4048763 0.5560494 0.5395652
#> [2976] 0.5608116 0.4005732 0.4039872 0.3985609 0.5897662 0.5236547 0.5294637
#> [2983] 0.5555485 0.2686676 0.6129223 0.6303603 0.6229623 0.5079629 0.5588753
#> [2990] 0.2465271 0.4056581 0.5990546 0.6055874 0.5484727 0.5684734 0.3977737
#> [2997] 0.6052874 0.5870685 0.6369145 0.3714917 0.3617024 0.4506418 0.3886282
#> [3004] 0.6345640 0.5099163 0.6035904 0.5571339 0.4133299 0.6290405 0.3908396
#> [3011] 0.5199217 0.5669060 0.4821393 0.4557733 0.6298858 0.5607906 0.4383318
#> [3018] 0.4594815 0.6002381 0.5865120 0.6124669 0.6284232 0.5671489 0.6284003
#> [3025] 0.5420971 0.6312175 0.5443381 0.5475840 0.6096808 0.4621868 0.5127396
#> [3032] 0.5023747 0.6240631 0.4405429 0.6358137 0.5611627 0.6112980 0.6223396
#> [3039] 0.6348745 0.2195769 0.5095285 0.2744031 0.6279543 0.4791295 0.4960339
#> [3046] 0.5726489 0.4870062 0.6252861 0.6217183 0.3224919 0.5416352 0.5986084
#> [3053] 0.5081307 0.5394116 0.5940569 0.5224773 0.5966001 0.3017163 0.3742333
#> [3060] 0.5910920 0.5926260 0.3108268 0.4880133 0.3651566 0.2838823 0.2789640
#> [3067] 0.5634282 0.3472466 0.5993304 0.5556464 0.5686327 0.3035810 0.5957469
#> [3074] 0.5594414 0.3396791 0.5740963 0.4157519 0.6025724 0.5079469 0.4356411
#> [3081] 0.2594919 0.5167535 0.5378989 0.6290381 0.6168472 0.6193743 0.5276553
#> [3088] 0.6338946 0.3479290 0.5301658 0.6360771 0.5862131 0.5814204 0.6334281
#> [3095] 0.4357117 0.4601636 0.5095293 0.3715081 0.6037014 0.6328119 0.5964084
#> [3102] 0.5123215 0.6331429 0.5971449 0.5681531 0.3454625 0.4296366 0.5661241
#> [3109] 0.3335101 0.6303662 0.4151469 0.5677394 0.3544808 0.4501487 0.4639171
#> [3116] 0.6034624 0.6338246 0.5328672 0.6278688 0.5794097 0.5072947 0.4090429
#> [3123] 0.5719483 0.4080424 0.6179278 0.5691656 0.5099912 0.6072236 0.5008272
#> [3130] 0.4621762 0.5526879 0.3515770 0.5803422 0.4991441 0.6203660 0.6177647
#> [3137] 0.5983856 0.4976037 0.4783445 0.4416166 0.4589907 0.5377233 0.6225346
#> [3144] 0.5976691 0.5910418 0.5601840 0.4608319 0.3674940 0.4488234 0.3917213
#> [3151] 0.3104710 0.5340118 0.4837892 0.3847200 0.6233539 0.3821499 0.4922634
#> [3158] 0.5625741 0.3053865 0.2059272 0.6138001 0.6296119 0.4190927 0.4377942
#> [3165] 0.3663282 0.6323003 0.5808398 0.5505158 0.5230135 0.3766010 0.6251148
#> [3172] 0.5659120 0.6064485 0.6155253 0.6291089 0.4067109 0.5380887 0.4609468
#> [3179] 0.4597309 0.3747536 0.5837643 0.5841121 0.3302312 0.2725686 0.6314354
#> [3186] 0.6278461 0.5580728 0.5006873 0.3118347 0.5884017 0.5890860 0.2866696
#> [3193] 0.6034136 0.5482136 0.6100360 0.5190659 0.6165502 0.4509209 0.3848811
#> [3200] 0.6152107 0.6307051 0.2237130 0.5985014 0.5513034 0.5334154 0.1803444
#> [3207] 0.4588482 0.4145498 0.5435004 0.3790794 0.4278706 0.6363431 0.5348789
#> [3214] 0.4457033 0.4776953 0.5582361 0.6280066 0.6267008 0.6297786 0.3352116
#> [3221] 0.6368384 0.4434086 0.2929111 0.4487645 0.4572662 0.4861861 0.5308566
#> [3228] 0.5742663 0.6273035 0.6125877 0.6345757 0.5228255 0.3958805 0.6302970
#> [3235] 0.5024217 0.5680390 0.5133716 0.5730094 0.4488147 0.6264545 0.6304787
#> [3242] 0.5562242 0.4158636 0.5986554 0.3875800 0.3328418 0.5796904 0.4597937
#> [3249] 0.5539272 0.6257195 0.4707279 0.6495793 0.4429396 0.3819943 0.2692290
#> [3256] 0.4328839 0.5053416 0.5177133 0.5302316 0.3344451 0.6363016 0.2036768
#> [3263] 0.6334453 0.6324899 0.3900184 0.3815351 0.3945682 0.6171502 0.4596167
#> [3270] 0.6118369 0.4756425 0.3579673 0.6301999 0.6302865 0.4723117 0.5450681
#> [3277] 0.6357869 0.6333323 0.6022677 0.4981844 0.5934648 0.2035852 0.4804533
#> [3284] 0.6267618 0.6307542 0.4338189 0.6369112 0.5105342 0.5846114 0.6304740
#> [3291] 0.2557048 0.6146276 0.2950682 0.6338530 0.3790833 0.6118522 0.5215659
#> [3298] 0.6013388 0.4584549 0.4614093 0.4965611 0.5895394 0.4908041 0.5638545
#> [3305] 0.4570528 0.4718187 0.5656123 0.4917134 0.6016131 0.5255681 0.6090994
#> [3312] 0.6276306 0.3197589 0.4975689 0.5196691 0.5945209 0.5995139 0.4732161
#> [3319] 0.6347630 0.6154321 0.5883487 0.3593984 0.3603013 0.6108356 0.5565188
#> [3326] 0.3504557 0.4637155 0.5644467 0.3916369 0.5309413 0.5541205 0.2761516
#> [3333] 0.5050663 0.3887957 0.6103004 0.2496128 0.6210868 0.5507960 0.4603478
#> [3340] 0.4644544 0.5505476 0.5351022 0.4396569 0.6119056 0.5840381 0.3821883
#> [3347] 0.5532926 0.6324973 0.3412061 0.6174933 0.5581975 0.4561430 0.1900342
#> [3354] 0.6109210 0.2664588 0.5785975 0.4593907 0.4720374 0.5789944 0.5740617
#> [3361] 0.2501548 0.4661383 0.2340146 0.5129852 0.4430325 0.6102555 0.5338155
#> [3368] 0.5626095 0.4302414 0.5053798 0.5969120 0.6265981 0.5409168 0.4743203
#> [3375] 0.6274947 0.5572615 0.3997274 0.5376808 0.4708863 0.3462822 0.3702093
#> [3382] 0.5243804 0.4324930 0.6068250 0.5142484 0.5146385 0.3595211 0.5439348
#> [3389] 0.5547316 0.6242055 0.5527910 0.4887319 0.6238439 0.5088921 0.4925699
#> [3396] 0.5258624 0.2644215 0.2959666 0.2390594 0.6214892 0.5131401 0.4367471
#> [3403] 0.5594030 0.4392869 0.3289065 0.4991783 0.5900264 0.5582312 0.5958696
#> [3410] 0.1783524 0.4638378 0.6085905 0.3491628 0.6369106 0.4482920 0.5533517
#> [3417] 0.6270786 0.6183351 0.4603823 0.6144142 0.5482006 0.4882810 0.5286802
#> [3424] 0.5688092 0.5257236 0.6312884 0.4028366 0.6307428 0.5058926 0.6369117
#> [3431] 0.4867315 0.6280862 0.4638770 0.4692417 0.4680600 0.4067304 0.6217433
#> [3438] 0.4947209 0.5732661 0.2212103 0.5292179 0.5768841 0.6142655 0.4843779
#> [3445] 0.6014931 0.6282535 0.5870400 0.5176399 0.2780195 0.5906367 0.5911331
#> [3452] 0.5119567 0.4438671 0.3201005 0.6360731 0.5878895 0.5072859 0.4329320
#> [3459] 0.6320781 0.3777064 0.6232556 0.6113989 0.6357420 0.5396175 0.5044311
#> [3466] 0.5035840 0.4840896 0.6353534 0.2919898 0.4033767 0.5203890 0.5718188
#> [3473] 0.5580543 0.6255816 0.6298438 0.4672684 0.4173982 0.5744075 0.3554071
#> [3480] 0.3801903 0.6010016 0.4203380 0.5689697 0.4333713 0.6134532 0.5824631
#> [3487] 0.6101269 0.6186814 0.5184272 0.4890491 0.4986742 0.6306463 0.6258287
#> [3494] 0.1934887 0.2711501 0.6087002 0.6143715 0.4514379 0.3784509 0.3706565
#> [3501] 0.3196433 0.4195021 0.5921564 0.5845498 0.4913533 0.6142648 0.4766353
#> [3508] 0.4109640 0.5154899 0.4209179 0.5454195 0.3837023 0.6210054 0.4722505
#> [3515] 0.6259796 0.6149687 0.5235772 0.5963871 0.5787429 0.3295699 0.6029252
#> [3522] 0.3187022 0.6157331 0.5978467 0.4609694 0.2972403 0.5722204 0.6081280
#> [3529] 0.5779518 0.6329917 0.5633119 0.6232546 0.4458382 0.5635674 0.6115757
#> [3536] 0.4872318 0.4030873 0.3909856 0.3636743 0.6360057 0.6187438 0.6059997
#> [3543] 0.6293559 0.6075114 0.5300763 0.3177626 0.6083272 0.3415715 0.6258765
#> [3550] 0.3821465 0.5211315 0.5429676 0.6245162 0.6100453 0.4443913 0.6001359
#> [3557] 0.5687252 0.4239925 0.5306718 0.5289823 0.5131618 0.5347994 0.4716853
#> [3564] 0.5273152 0.6190847 0.3692698 0.4158002 0.6147544 0.5709565 0.4924391
#> [3571] 0.4738944 0.3931239 0.6262778 0.4071578 0.3523648 0.5916655 0.5600956
#> [3578] 0.5039561 0.6102051 0.5691568 0.6267862 0.5346640 0.2540282 0.5807614
#> [3585] 0.6263514 0.6249909 0.4986141 0.3438650 0.6186375 0.4323231 0.4551286
#> [3592] 0.3688901 0.5478050 0.6247936 0.3984848 0.3794920 0.6281247 0.5527842
#> [3599] 0.5204505 0.6369086 0.2888141 0.5180764 0.5982488 0.5068124 0.5512984
#> [3606] 0.5676704 0.5173552 0.4727647 0.4305215 0.5690880 0.5739830 0.5130377
#> [3613] 0.6093084 0.5640937 0.5193017 0.5142176 0.4586762 0.4122244 0.1951309
#> [3620] 0.5673822 0.6024857 0.6076701 0.6297671 0.6247530 0.3913186 0.4630804
#> [3627] 0.3647691 0.6274505 0.5203262 0.4477268 0.3184938 0.5659710 0.6316619
#> [3634] 0.4527461 0.5012234 0.5209883 0.3635174 0.4124998 0.5881357 0.3273297
#> [3641] 0.6145006 0.5025666 0.5238870 0.6008323 0.4254784 0.3009745 0.4344746
#> [3648] 0.3203702 0.6105185 0.2781995 0.4907878 0.5371477 0.4828769 0.4141057
#> [3655] 0.3591675 0.4600314 0.5740461 0.6317733 0.6264694 0.3522694 0.5435439
#> [3662] 0.5189347 0.6349704 0.5190597 0.6269514 0.4769063 0.5008457 0.6024170
#> [3669] 0.6292064 0.5598544 0.6025701 0.6185829 0.5750393 0.4402956 0.6183090
#> [3676] 0.3596946 0.4837734 0.6278878 0.6338061 0.5923131 0.6268727 0.6263028
#> [3683] 0.4567942 0.6359031 0.6150230 0.3624116 0.5677808 0.5826796 0.5461830
#> [3690] 0.2898668 0.4027548 0.3015389 0.5658301 0.3826520 0.3229918 0.6184371
#> [3697] 0.3390099 0.6179911 0.5653373 0.5120229 0.3063482 0.4881999 0.6114410
#> [3704] 0.4998808 0.5148229 0.4683375 0.5273384 0.5112149 0.5181652 0.6350421
#> [3711] 0.5710861 0.6308949 0.5900232 0.5454901 0.4417450 0.6369106 0.4792824
#> [3718] 0.6361796 0.6182973 0.2573387 0.5486135 0.6298308 0.6305673 0.4891774
#> [3725] 0.4490347 0.6334593 0.3764663 0.5461177 0.3771629 0.4458399 0.5103517
#> [3732] 0.5418205 0.3940085 0.5972542 0.6135480 0.6110561 0.4672153 0.6369102
#> [3739] 0.2304307 0.3838445 0.4786765 0.4425763 0.4701290 0.5338481 0.6160972
#> [3746] 0.5732186 0.5323334 0.4362384 0.6291748 0.6339490 0.6257003 0.2807371
#> [3753] 0.6212660 0.6359416 0.3601384 0.2391376 0.5986597 0.4068304 0.5065448
#> [3760] 0.5262056 0.5464248 0.6348024 0.5064300 0.6104051 0.6093447 0.6136183
#> [3767] 0.4458391 0.5467007 0.6332654 0.4860483 0.5001805 0.4086992 0.5221526
#> [3774] 0.5055404 0.2713342 0.4620180 0.5919783 0.4765556 0.4461742 0.5328085
#> [3781] 0.5046388 0.3940998 0.5412305 0.4774405 0.6027331 0.3607022 0.6151030
#> [3788] 0.2887006 0.3381159 0.6151653 0.4549825 0.6363679 0.4510402 0.6257574
#> [3795] 0.6227050 0.5219282 0.6360436 0.6246924 0.3346736 0.3931921 0.6121317
#> [3802] 0.5879666 0.4815791 0.6269361 0.5744716 0.5015397 0.6219134 0.6074028
#> [3809] 0.5458408 0.6042861 0.1949182 0.6344201 0.4285268 0.4021498 0.5459987
#> [3816] 0.5440071 0.5167993 0.5725012 0.6346676 0.4528055 0.5354914 0.6313299
#> [3823] 0.5533930 0.6320343 0.5346752 0.3778646 0.3983797 0.4053741 0.5454896
#> [3830] 0.3564750 0.6317236 0.5235996 0.4143179 0.6146272 0.6330911 0.5218608
#> [3837] 0.4810961 0.4951412 0.3472468 0.5714967 0.4426743 0.4485942 0.5669155
#> [3844] 0.3379695 0.4388100 0.6265951 0.5072967 0.6341886 0.5310442 0.4546522
#> [3851] 0.5670802 0.5880021 0.4453740 0.3691828 0.3198036 0.4542413 0.4507054
#> [3858] 0.4120320 0.5230748 0.5833598 0.4710414 0.5990232 0.6182452 0.4936832
#> [3865] 0.4132809 0.3188285 0.5606089 0.3754222 0.3316384 0.2239015 0.4997745
#> [3872] 0.6039282 0.3664721 0.3172271 0.3389916 0.5709806 0.5308513 0.6143161
#> [3879] 0.4797094 0.5480376 0.6173360 0.5807122 0.4503719 0.5936099 0.4976486
#> [3886] 0.6036774 0.4701907 0.5367710 0.5135639 0.5423534 0.3282334 0.6177942
#> [3893] 0.4372130 0.6291099 0.4599810 0.5885721 0.4640143 0.6074936 0.4386896
#> [3900] 0.6097506 0.4728907 0.4274533 0.5730902 0.5804475 0.6082078 0.3463848
#> [3907] 0.2280309 0.5159451 0.5207174 0.3562619 0.4189675 0.5623568 0.4752830
#> [3914] 0.5372956 0.4442139 0.1096242 0.5952247 0.3999832 0.4708116 0.4906764
#> [3921] 0.5241325 0.4235113 0.3216144 0.5868983 0.2373504 0.6036974 0.5214860
#> [3928] 0.6338367 0.6295658 0.5951055 0.5980147 0.4561856 0.4785608 0.4656968
#> [3935] 0.4286087 0.5705288 0.4513459 0.3284322 0.5732264 0.5934668 0.5393466
#> [3942] 0.6333448 0.5270635 0.5811503 0.6368463 0.6121571 0.5926069 0.5306942
#> [3949] 0.5022635 0.6348287 0.6255405 0.5106743 0.6355242 0.5975021 0.5681874
#> [3956] 0.4883251 0.6029598 0.6337858 0.5271924 0.3306315 0.5387904 0.6368013
#> [3963] 0.5721811 0.5902605 0.4411248 0.5619797 0.5831648 0.2448906 0.3805973
#> [3970] 0.3331106 0.4781760 0.6091538 0.3075213 0.5210539 0.4882983 0.4362384
#> [3977] 0.3286332 0.4117142 0.6365419 0.5811538 0.3618724 0.5165962 0.6155566
#> [3984] 0.5615394 0.6096728 0.6213170 0.3911546 0.5708003 0.6216691 0.4205463
#> [3991] 0.4111180 0.4160102 0.3819556 0.6206485 0.6264400 0.5383150 0.3897398
#> [3998] 0.4991718 0.5197969 0.5634302 0.5210519 0.5040387 0.6293950 0.5012811
#> [4005] 0.4871134 0.5995844 0.5744858 0.4879010 0.5109101 0.6214431 0.5885454
#> [4012] 0.6364190 0.3569088 0.3303843 0.6298271 0.5987590 0.4595496 0.4577400
#> [4019] 0.5375422 0.6022653 0.5471628 0.6157577 0.5631214 0.5489576 0.5854843
#> [4026] 0.6146872 0.5993336 0.5354352 0.6313256 0.4850182 0.6092623 0.3824842
#> [4033] 0.4153150 0.6301103 0.4463964 0.6368904 0.5619797 0.5930009 0.1891379
#> [4040] 0.6159425 0.5882557 0.4971581 0.6272486 0.4797451 0.4364980 0.2852417
#> [4047] 0.6112165 0.6097144 0.3426063 0.3940487 0.3671545 0.2281678 0.5540274
#> [4054] 0.5569929 0.6210620 0.5320490 0.6186190 0.5088421 0.3541914 0.5356375
#> [4061] 0.4103359 0.6194010 0.5990995 0.3088515 0.5364505 0.4909019 0.6007837
#> [4068] 0.4947507 0.6000907 0.6016565 0.3720277 0.5319102 0.3095872 0.3828208
#> [4075] 0.3868484 0.6253720 0.6213589 0.3877948 0.5769803 0.1563225 0.3823040
#> [4082] 0.2574433 0.2584788 0.5559601 0.6310590 0.6117479 0.4054532 0.6281009
#> [4089] 0.3217017 0.5700503 0.6095712 0.5057166 0.6362077 0.5821489 0.5351064
#> [4096] 0.5265292 0.6242222 0.4471758 0.3239184 0.6083604 0.6331865 0.4133356
#> [4103] 0.5790284 0.5542567 0.2706862 0.3517855 0.4602144 0.5858595 0.4963304
#> [4110] 0.5350697 0.5185547 0.5362855 0.6105633 0.5506644 0.6070822 0.5905025
#> [4117] 0.3400886 0.5127954 0.6184899 0.6352624 0.6214443 0.5350407 0.4190501
#> [4124] 0.5100468 0.4192089 0.5210347 0.6254598 0.3626406 0.4579501 0.5548118
#> [4131] 0.6245711 0.6369106 0.4491597 0.6152400 0.5827681 0.4154780 0.6313749
#> [4138] 0.6033329 0.5402672 0.4852744 0.5821671 0.4409336 0.5330259 0.3593879
#> [4145] 0.6325159 0.3559700 0.5669812 0.3657824 0.6315020 0.6033310 0.6289124
#> [4152] 0.4489787 0.3434387 0.5297727 0.5211333 0.4586662 0.5610688 0.5386223
#> [4159] 0.6146293 0.5708801 0.2067429 0.3633706 0.6358211 0.4669696 0.6149925
#> [4166] 0.4484755 0.3863297 0.3619725 0.6369106 0.6221479 0.6020897 0.3358729
#> [4173] 0.5948720 0.5798501 0.2406030 0.3974331 0.6079447 0.5226196 0.5688166
#> [4180] 0.5905531 0.2499876 0.4897748 0.6314152 0.5682291 0.5005882 0.5898920
#> [4187] 0.5852744 0.6142349 0.5592701 0.5126985 0.4888525 0.5631480 0.3253444
#> [4194] 0.5005871 0.5238483 0.5886658 0.1757015 0.5821659 0.6115069 0.2789167
#> [4201] 0.6234444 0.5963998 0.5315794 0.3556426 0.6349985 0.6231765 0.4628741
#> [4208] 0.5352997 0.4345677 0.6173525 0.3958366 0.5437832 0.5319007 0.6327879
#> [4215] 0.3799331 0.3649243 0.3455730 0.5361298 0.6025531 0.5849094 0.6233077
#> [4222] 0.5879170 0.3787164 0.5570869 0.3864203 0.5231225 0.4420124 0.6089305
#> [4229] 0.5915629 0.4611330 0.5461340 0.3948583 0.6271072 0.6299048 0.4518073
#> [4236] 0.5324884 0.3593800 0.6225875 0.4667405 0.5125028 0.6116714 0.6114348
#> [4243] 0.6044332 0.3782661 0.5953791 0.6345287 0.5888894 0.6366901 0.5159317
#> [4250] 0.6150809 0.4696732 0.4777463 0.5189878 0.4536016 0.4584944 0.5126900
#> [4257] 0.5193451 0.3611183 0.6091416 0.3472564 0.3719584 0.4319101 0.6357095
#> [4264] 0.6333032 0.4653752 0.4400601 0.4621183 0.4426162 0.5661898 0.5839866
#> [4271] 0.2645765 0.6324547 0.5594476 0.5191577 0.5995430 0.6167255 0.6301343
#> [4278] 0.3524658 0.4655258 0.6077885 0.6091171 0.6223702 0.5859277 0.4272668
#> [4285] 0.4019224 0.3721495 0.3183092 0.5095277 0.5667522 0.5270506 0.5940202
#> [4292] 0.3186357 0.4287824 0.5165985 0.5190420 0.4599610 0.5566849 0.4905402
#> [4299] 0.3971028 0.4803544 0.6310353 0.5494902 0.6298573 0.2492812 0.5510196
#> [4306] 0.4532847 0.5763562 0.5491684 0.6313878 0.5619802 0.3931553 0.5650549
#> [4313] 0.3122940 0.5692971 0.5984222 0.5319235 0.4091916 0.3653561 0.4828112
#> [4320] 0.6251863 0.5925445 0.4778107 0.4343600 0.3202162 0.4033376 0.5844927
#> [4327] 0.6365427 0.5905603 0.3286733 0.5488324 0.4756731 0.6112873 0.5458520
#> [4334] 0.5400712 0.5455378 0.5113755 0.6216147 0.4718333 0.5960861 0.5579662
#> [4341] 0.5280775 0.5785965 0.5354532 0.5313242 0.6315661 0.4114966 0.5460248
#> [4348] 0.3384057 0.3755427 0.6293490 0.4234694 0.5346210 0.4891615 0.5293833
#> [4355] 0.3134098 0.3298352 0.4334266 0.3053160 0.4333696 0.4498249 0.2942889
#> [4362] 0.5267424 0.4718064 0.3426198 0.6175684 0.4980750 0.6028395 0.4555448
#> [4369] 0.5794514 0.6230575 0.6140911 0.5940654 0.6164818 0.5969333 0.4840193
#> [4376] 0.5507045 0.4063337 0.5953641 0.2969890 0.6296131 0.6339095 0.6316090
#> [4383] 0.4610859 0.5707875 0.5960526 0.5532026 0.5864438 0.5058502 0.5067188
#> [4390] 0.4656487 0.3934297 0.4464434 0.6126914 0.5586972 0.4549774 0.5205904
#> [4397] 0.5917041 0.5376045 0.6115122 0.6137490 0.4728770 0.6367022 0.6354970
#> [4404] 0.5907463 0.5212008 0.1453867 0.3099957 0.4632644 0.6153668 0.4882820
#> [4411] 0.5779250 0.3791398 0.3276905 0.2058800 0.6148391 0.5804731 0.5633594
#> [4418] 0.6062137 0.5431351 0.5378980 0.5741983 0.4739580 0.5654598 0.3985540
#> [4425] 0.3648369 0.5797361 0.6302228 0.5637265 0.5317501 0.5631197 0.6249316
#> [4432] 0.3670435 0.5421665 0.5177145 0.6150289 0.5506808 0.6158615 0.6363551
#> [4439] 0.5011015 0.6174235 0.6170199 0.3096611 0.4519752 0.6253330 0.5339525
#> [4446] 0.6227528 0.3997533 0.4141252 0.6303104 0.6356039 0.5432019 0.3171329
#> [4453] 0.4081391 0.4183592 0.6346883 0.6242866 0.6357510 0.4238847 0.4617965
#> [4460] 0.4594035 0.4140467 0.1655853 0.3882195 0.4241209 0.4714462 0.5742087
#> [4467] 0.4309715 0.5563638 0.5703033 0.3659860 0.5588230 0.4385503 0.5708888
#> [4474] 0.6321887 0.5283159 0.6262963 0.6319193 0.4950131 0.3416201 0.5309657
#> [4481] 0.5743674 0.6292675 0.4299350 0.4132151 0.5252471 0.6360271 0.5352784
#> [4488] 0.4776243 0.3584114 0.6347979 0.4082163 0.6320740 0.4619120 0.4345062
#> [4495] 0.3689240 0.3254400 0.2629398 0.6329414 0.6178299 0.6355908 0.4694995
#> [4502] 0.3577770 0.4467328 0.6316813 0.4552549 0.3014941 0.3506316 0.5509595
#> [4509] 0.6165754 0.6155829 0.3952426 0.6231411 0.4941218 0.4262298 0.5909181
#> [4516] 0.6275145 0.3327607 0.2413231 0.5083997 0.5487230 0.3510118 0.5847276
#> [4523] 0.3109435 0.6236039 0.5061212 0.6203271 0.4380845 0.5102531 0.4588599
#> [4530] 0.4352797 0.5615087 0.5676219 0.5649160 0.5573477 0.6164411 0.6157224
#> [4537] 0.4667246 0.1955740 0.5856601 0.6099131 0.6133317 0.5434854 0.5968405
#> [4544] 0.2203402 0.4824638 0.6181267 0.2903717 0.4198612 0.4820502 0.5962295
#> [4551] 0.4932247 0.2393185 0.4020831 0.4982674 0.6057864 0.5034477 0.6279723
#> [4558] 0.2713626 0.5424771 0.6195944 0.6105145 0.3590346 0.5005374 0.5442365
#> [4565] 0.6204369 0.6278050 0.5157551 0.5195190 0.4862176 0.5613418 0.6348135
#> [4572] 0.5259155 0.6052678 0.6311524 0.4202718 0.4789861 0.4021168 0.4907479
#> [4579] 0.5068072 0.4774207 0.3757992 0.6309779 0.4676612 0.6339775 0.6156573
#> [4586] 0.6313328 0.5741791 0.5496476 0.5669002 0.5787881 0.6360279 0.4448745
#> [4593] 0.5080987 0.4768573 0.3220227 0.4328498 0.5398282 0.4244484 0.4984423
#> [4600] 0.6403123 0.4432516 0.6304975 0.6029341 0.6167257 0.4375318 0.3654554
#> [4607] 0.4385043 0.5980246 0.3912157 0.5945868 0.4851988 0.3245049 0.4874066
#> [4614] 0.4018720 0.5553939 0.5500917 0.6129676 0.4642407 0.5101783 0.4742019
#> [4621] 0.6219368 0.6026100 0.6321751 0.4695805 0.4160222 0.5568298 0.6113268
#> [4628] 0.6275907 0.3325223 0.4086579 0.5625742 0.5788110 0.4819513 0.6060625
#> [4635] 0.6131673 0.6287678 0.5920882 0.5902790 0.3348093 0.5930340 0.4435125
#> [4642] 0.6325741 0.5520966 0.3211748 0.1983036 0.5341876 0.2085808 0.4727417
#> [4649] 0.3336638 0.3900748 0.5214936 0.5706710 0.5013279 0.3858841 0.4766790
#> [4656] 0.6126394 0.4846971 0.5788290 0.5629782 0.5684513 0.6294643 0.4517216
#> [4663] 0.5416332 0.6209928 0.6234923 0.4370411 0.6141623 0.6227825 0.6302014
#> [4670] 0.6323689 0.4290763 0.6347432 0.6317379 0.5784941 0.6336012 0.6191450
#> [4677] 0.6293329 0.5522998 0.6335423 0.5039744 0.5731648 0.3763982 0.4044633
#> [4684] 0.5580479 0.5856989 0.6026175 0.6234935 0.4355524 0.5003682 0.5378493
#> [4691] 0.5577187 0.6307673 0.5808799 0.3715707 0.5496894 0.3419283 0.3590336
#> [4698] 0.5013951 0.5741606 0.5343947 0.6021085 0.6279210 0.6115905 0.4816849
#> [4705] 0.4895301 0.6288869 0.4722086 0.4253644 0.6293551 0.4861135 0.5725141
#> [4712] 0.6128914 0.5499568 0.5766701 0.5737479 0.6013911 0.6201410 0.4055715
#> [4719] 0.6103754 0.3503273 0.5079901 0.6256653 0.5440048 0.5223941 0.6327004
#> [4726] 0.6297459 0.5307329 0.5378900 0.4627012 0.4004260 0.6346723 0.5944663
#> [4733] 0.5505734 0.3934001 0.6132112 0.4062199 0.3738838 0.2694507 0.2024100
#> [4740] 0.3773575 0.3970910 0.5674371 0.5168236 0.5356884 0.5632947 0.2549956
#> [4747] 0.5893061 0.4354320 0.5173752 0.5339524 0.3419737 0.4814351 0.5846620
#> [4754] 0.4827988 0.3627688 0.6336914 0.6265528 0.4631218 0.5312448 0.5600759
#> [4761] 0.5098199 0.6259500 0.2941946 0.4874663 0.4620313 0.4456736 0.6235104
#> [4768] 0.5985780 0.3608285 0.4104356 0.5682689 0.5566831 0.5499328 0.6050595
#> [4775] 0.6291885 0.5017432 0.4141900 0.4269333 0.2958799 0.3414143 0.5880652
#> [4782] 0.6352995 0.5762762 0.3546546 0.5179475 0.3122736 0.4354972 0.2533707
#> [4789] 0.5122746 0.5035657 0.5989641 0.4978931 0.4229810 0.6090158 0.5080259
#> [4796] 0.6292715 0.6023587 0.5267277 0.3833753 0.5481163 0.4867613 0.4449092
#> [4803] 0.4936849 0.4781572 0.5758537 0.5427582 0.5534020 0.6247435 0.3017951
#> [4810] 0.5540690 0.6078565 0.5394593 0.4751266 0.4510126 0.3561680 0.6199954
#> [4817] 0.6353841 0.5369593 0.3569629 0.5326531 0.5731053 0.2659282 0.5882816
#> [4824] 0.3637888 0.4868961 0.4664551 0.3748456 0.6262747 0.5601560 0.4942092
#> [4831] 0.6097262 0.4618201 0.5521196 0.5914855 0.4951327 0.5597387 0.4215587
#> [4838] 0.3279098 0.4866199 0.5289028

# how the calculation is done - first observation
fs_area(buildings[1,]) / fs_area(fs_mbc(buildings[1,]))

#> 0.573271 [1]